### 2023-10-02 #5921 @editorialbot
[PRE REVIEW]: stemflow: A Python Package for Adaptive Spatio-Temporal Exploratory Model
**Submitting author:** @chenyangkang (<a href="http://orcid.org/0000-0003-4996-0881">Yangkang Chen</a>)
**Repository:** https://github.com/chenyangkang/stemflow
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
