### 2023-10-08 426b3d9
update plot gif
