+++ docs/API/stemflow.model.special_hurdle.md
@@ -1,5 +1,0 @@
-# stemflow.model.special_hurdle
-
----
-:::stemflow.model.special_hurdle
----
