+++ docs/API/stemflow.utils.plot_gif.md
@@ -1,5 +1,0 @@
-# stemflow.utils.plot_gif
-
----
-:::stemflow.utils.plot_gif
----
