+++ docs/API/stemflow.utils.quadtree.md
@@ -1,5 +1,0 @@
-# stemflow.utils.quadtree
-
----
-:::stemflow.utils.quadtree
----
