+++ docs/API/stemflow.model.dummy_model.md
@@ -1,5 +1,0 @@
-# stemflow.model.dummy_model
-
----
-:::stemflow.model.dummy_model
----
