+++ docs/API/stemflow.utils.generate_soft_colors.md
@@ -1,3 +1,0 @@
-# stemflow.utils.generate_soft_colors
-
-:::stemflow.utils.generate_soft_colors
