+++ docs/Fun_Visualization/Global_Bird_Shannon_H_Index.md
@@ -1,0 +1,9 @@
+# Global Bird Diversity (Shannon H) Index
+
+Data Source: eBird
+
+
+![](./shannon_H_2020.gif)
+
+
+
