+++ docs/API/stemflow.model.Hurdle.md
@@ -1,5 +1,0 @@
-# stemflow.model.Hurdle
-
----
-:::stemflow.model.Hurdle
----
