+++ docs/API/stemflow.model_selection.md
@@ -1,3 +1,0 @@
-# stemflow.model_selection
-
-::: stemflow.model_selection
