+++ docs/Fun_Visualization/Global_NDVI.md
@@ -1,0 +1,9 @@
+# Global NDVI (Normalized Difference Vegetation Index)
+
+Data Source: [MODIS VI products (MOD13)](https://modis.gsfc.nasa.gov/data/dataprod/mod13.php)
+
+
+![](./NDVI_2020.gif)
+
+
+
