+++ docs/Fun_Visualization/Ruby-crowned_Kinglet_2020.md
@@ -1,0 +1,9 @@
+# Seasonal Distribution of Ruby-crowned_Kinglet
+
+Data Source: eBird
+
+
+![](./Ruby-crowned_Kinglet_2020.gif)
+
+It's quite amazing that the Ruby-crowned_Kinglet distribution shows two distinct populations that maintain a non-breeding site fidelity instead of breeding site fidelity.
+
