### 2023-10-17 729a9a4
No x_train copy while training
