### 2023-10-17 48859cb
change stixel requesting way when training to make it more efficient
