### 2023-10-18 97169cb
add multiprocessing while spliting and generating stixels
