### 2023-10-18 8d5073a
fix multiprocess parameter problems while training
