+++ stemflow/mini_test.py
@@ -18,7 +18,7 @@ import h3pandas
 
 # warnings.filterwarnings('ignore')
 
-def run_mini_test(delet_tmp_files: bool=True, show: bool = False):
+def run_mini_test(delet_tmp_files: bool=True, show: bool = False, ensemble_models_disk_saver=False, ensemble_models_disk_saving_dir='./'):
     """Run a mini test
     
     Processes:
@@ -89,6 +89,11 @@ def run_mini_test(delet_tmp_files: bool=True, show: bool = False):
     # # First thing first: Spatio-temporal train test split
 
     # %%
+    import stemflow
+    isl_path = stemflow.__path__[0]
+    print(f'Installation path: {isl_path}')
+    
+    #
     print('ST_train_test_split ...')
     from stemflow.model_selection import ST_train_test_split
     X_train, X_test, y_train, y_test = ST_train_test_split(X, y, 
@@ -108,42 +113,33 @@ def run_mini_test(delet_tmp_files: bool=True, show: bool = False):
 
     # %%
     print('Declaring model instance...')
-    model = Hurdle_for_AdaSTEM(
-        classifier=AdaSTEMClassifier(base_model=XGBClassifier(tree_method='hist',random_state=42, verbosity = 0, n_jobs=1),
-                                    save_gridding_plot = True,
-                                    ensemble_fold=5, 
-                                    min_ensemble_required=3,
-                                    grid_len_lon_upper_threshold=10,
-                                    grid_len_lon_lower_threshold=2,
-                                    grid_len_lat_upper_threshold=10,
-                                    grid_len_lat_lower_threshold=2,
-                                    points_lower_threshold=50,
-                                    Spatio1='longitude',
-                                    Spatio2 = 'latitude', 
-                                    Temporal1 = 'DOY',
-                                    use_temporal_to_train=True,
-                                    njobs=1,
-                                    plot_xlims=(data.longitude.min(), data.longitude.max()), 
-                                    plot_ylims=(data.latitude.min(),data.latitude.max())
-                                    ),
-        regressor=AdaSTEMRegressor(base_model=XGBRegressor(tree_method='hist',random_state=42, verbosity = 0, n_jobs=1),
-                                    save_gridding_plot = True,
-                                    ensemble_fold=5, 
-                                    min_ensemble_required=3,
-                                    grid_len_lon_upper_threshold=10,
-                                    grid_len_lon_lower_threshold=2,
-                                    grid_len_lat_upper_threshold=10,
-                                    grid_len_lat_lower_threshold=2,
-                                    points_lower_threshold=20,
-                                    Spatio1='longitude',
-                                    Spatio2 = 'latitude', 
-                                    Temporal1 = 'DOY',
-                                    use_temporal_to_train=True,
-                                    njobs=1,
-                                    plot_xlims=(data.longitude.min(), data.longitude.max()), 
-                                    plot_ylims=(data.latitude.min(),data.latitude.max())
-                                    )
+    
+    model = AdaSTEMRegressor(
+        base_model=Hurdle(
+            classifier=XGBClassifier(tree_method='hist',random_state=42, verbosity = 0, n_jobs=1),
+            regressor=XGBRegressor(tree_method='hist',random_state=42, verbosity = 0, n_jobs=1)
+        ),
+        save_gridding_plot = True,
+        ensemble_fold=5, 
+        min_ensemble_required=3,
+        grid_len_lon_upper_threshold=10,
+        grid_len_lon_lower_threshold=2,
+        grid_len_lat_upper_threshold=10,
+        grid_len_lat_lower_threshold=2,
[…]
