+++ docs/index.md
@@ -208,6 +208,8 @@ Here, each color shows an ensemble generated during model fitting. In each of th
 ---- 
 ## Example of visualization :world_map:
 
+Daily Abundance Map of Barn Swallow
+
 ![GIF visualization](https://github.com/chenyangkang/stemflow/raw/main/docs/pred_gif.gif)
 
 See section [Prediction and Visualization](https://chenyangkang.github.io/stemflow/Examples/04.Prediction_visualization.html) for how to generate this GIF.
