### 2023-10-25 24b6168
fix quadtree for subdivision rules
