### 2023-11-03 faf5bc6
change spatial jitter scale
