### 2023-11-04 d562a3d
fix multiprocessing of split and fit
