### 2023-11-04 f9a1607
return self after fit
