### 2023-11-04 c526545
update mini_test
