+++ stemflow/mini_test.py
@@ -184,14 +184,14 @@ def run_mini_test(delet_tmp_files: bool=True, show: bool = False, ensemble_model
         ensemble_fold=fold_, 
         min_ensemble_required=min_req,
         grid_len_lon_upper_threshold=40,
-        grid_len_lon_lower_threshold=10,
-        grid_len_lat_upper_threshold=10,
-        grid_len_lat_lower_threshold=40,
+        grid_len_lon_lower_threshold=5,
+        grid_len_lat_upper_threshold=40,
+        grid_len_lat_lower_threshold=5,
         temporal_start=1, 
         temporal_end=366,
         temporal_step=30,
-        temporal_bin_interval=50,
-        points_lower_threshold=50,
+        temporal_bin_interval=60,
+        points_lower_threshold=30,
         Spatio1='longitude',
         Spatio2 = 'latitude', 
         Temporal1 = 'DOY',
