### 2023-12-14 #5921 @diazrenata
Hi @jedalong, would you be available and interested in reviewing this submission to JOSS?
