### 2023-12-14 #5921 @diazrenata
Hi @earth-chris, would you be available and interested in reviewing this submission to JOSS?
