### 2023-12-18 #5921 @diazrenata
@earth-chris Fantastic - thank you very much! I will add you as a reviewer and initiate the review process (which will open a separate issue).

@earth-chris and @jedalong - [JOSS's guide for reviewers](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) has a complete explanation of the review process and critera. The editorialbot system will generate checklists for you to help with reviewing, as well.

Thank you both for your time and attention!
