+++ docs/Fun_Visualization/Ruby-crowned_Kinglet_2020.md
@@ -5,5 +5,6 @@ Data Source: eBird
 
 ![](./Ruby-crowned_Kinglet_2020.gif)
 
-It's quite amazing that the Ruby-crowned_Kinglet distribution shows two distinct populations that maintain a non-breeding site fidelity instead of breeding site fidelity.
+It's quite amazing that the Ruby-crowned_Kinglet gathered for wintering and spread into two distinct populations during the breeding season.
+
 
