### 2023-12-21 1937da9
change spatial jitter magnitude to 100
