### 2024-01-03 d1071a6
add temperature visualization
