+++ docs/Fun_Visualization/Global_Mean_Temperature_2020.md
@@ -1,0 +1,10 @@
+# Seasonal Distribution of Ruby-crowned_Kinglet
+
+Data Source: ERA5
+
+
+![](./Temp_2020.gif)
+
+
+
+
