### 2024-01-05 #6158 @diazrenata
Hi @jedbrown & @earth-chris - happy new year! As you're ready, the review for this contribution will take place in this issue thread. The reviewer guidelines are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Thank you for your time!
