### 2024-01-08 chenyangkang/stemflow#3 @earth-chris
[REVIEW] consider adding automated testing
Automated testing is a best practice for open source projects, reducing the risk of software regression through breaking changes. `stemflow` includes a mini_test module that appears to test many parts of the software, but this is run manually. I recommend adding automated tests in CI to verify (and to provide verification to users) that the software continues to function as expected.
