### 2024-01-08 chenyangkang/stemflow#4 @earth-chris
[REVIEW] consider updating contributor guidelines
The `stemflow` README contains a sentence on welcoming contributions, but does not specify how those contributions are best made. There are some [nice guides](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) available for providing a contributor guide, and I recommend adding a more detailed guide for users.
