### 2024-01-08 #6158 @earth-chris
Hi @chenyangkang, I have concluded my review - thank you for the contribution! This is a valuable software package.

This software passes the majority of the checklist-based review guidelines, and the majority of my comments are related to improved documentation regarding the functionality of the package (see [here](https://github.com/chenyangkang/stemflow/issues/2)).

I also created new issues regarding [automated testing](https://github.com/chenyangkang/stemflow/issues/3) and [community guidelines](https://github.com/chenyangkang/stemflow/issues/4).

I don't believe any of the above issues merit withholding publication, and would be willing to check the boxes if @diazrenata is sufficiently satisfied (there are manual, not automated tests; a statement on contribution exists, but is minimal). But I at least wanted to flag them as minor issues.
