+++ docs/CONTRIBUTING.md
@@ -1,29 +1,0 @@
-# Contributing to stemflow
-
-We welcome contribution to `stemflow`
-
----
-
-## Steps for contributing
-
-1. Fork the git repository
-1. Create a development environment & install dependencies
-1. Create a new branch, make changes to code & add tests
-1. Update the docs
-1. Submit a pull request
-
----
-
-## Fork the git repository
-
-You will need your own fork to work on the code. Go to the `stemflow` project page and hit the Fork button. You will want to clone your fork to your machine:
-
-```bash
-git clone git@github.com:YOUR-USER-NAME/stemflow.git stemflow-YOURNAME
-cd stemflow-YOURNAME
-git remote add upstream git://github.com/stemflow/stemflow.git
-```
-
-This creates the directory `stemflow-YOURNAME` and connects your repository to the upstream (main project) elapid repository.
-
----
