### 2024-01-08 #6158 @diazrenata
Hi @jedalong - this issue is where review for the `stemflow` submission to JOSS will take place. As you're ready, the reviewer guidelines are available [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
