+++ docs/Fun_Visualization/Global_Mean_Temperature_2020.md
@@ -2,6 +2,8 @@
 
 Data Source: ERA5
 
+Notice: the visualized data is not modeled by AdaSTEM. Only visualization function is used here.
+
 
 ![](./Temp_2020.gif)
 
