+++ docs/Fun_Visualization/Global_NDVI.md
@@ -2,6 +2,8 @@
 
 Data Source: [MODIS VI products (MOD13)](https://modis.gsfc.nasa.gov/data/dataprod/mod13.php)
 
+Notice: the visualized data is not modeled by AdaSTEM. Only visualization function is used here.
+
 
 ![](./NDVI_2020.gif)
 
