+++ mini_test.py
@@ -1,299 +1,0 @@
-# %% [markdown]
-# # A **mini** test of stemflow
-# 
-# Yangkang Chen<br>
-# Sep 12, 2023
-
-# %%
-import pandas as pd
-import numpy as np
-import random
-from tqdm.auto import tqdm
-import matplotlib.pyplot as plt
-import matplotlib
-import warnings
-import pickle
-import os
-import h3pandas
-
-# warnings.filterwarnings('ignore')
-
-def run_mini_test(delet_tmp_files: bool=True, show: bool = False):
-    """Run a mini test
-    
-    Processes:
-        1. Request data
-        2. ST_train_test_split
-        3. Import stemflow modules
-        4. Declare model instance
-        5. Fitting model
-        6. Calculating feature importances
-        7. Assigning importance to points
-        8. Ploting top 2 important variables
-        9. Caclulating the fitting errors
-        10. Predicting on test set
-        11. Evaluation
-        12. Watermark
-        13. Deleting tmp files (optional)
-        
-    """
-    #
-    print('Start Runing Mini-test...')
-    print(f'Temporary files will be stored at: ./stemflow_mini_test/')
-    if delet_tmp_files:
-        print('Temporary files will be deleted.')
-    else:
-        print('Temporary files will *NOT* be deleted.')
-    # download mini data
-    if not os.path.exists('./stemflow_mini_test'):
-        os.makedirs('./stemflow_mini_test')
-    if not 'mini_data.csv' in os.listdir('./stemflow_mini_test'):
-        url = "https://chenyangkang.github.io/stemflow/mini_data/mini_data.csv"
-        print(f'Requesting data from {url} ...')
-        data = pd.read_csv(url)
-        data.to_csv('./stemflow_mini_test/mini_data.csv', index=False)
-        print('Done.')
-    else:
-        print('Mini-data already downloaded.')
-
-
-    # %%
-    # load data
-    data = pd.read_csv('./stemflow_mini_test/mini_data.csv').sample(frac=0.3, replace=False)
-
-
-    # %%
-    # data.head()
-
-    # %%
-    plt.scatter(
-        data.longitude,
-        data.latitude,
-        s=0.2
-    )
-    plt.savefig('./stemflow_mini_test/data_plot.pdf')
-    if show:
-        plt.show()
-    else:
-        plt.close()
-
-    # %% [markdown]
-    # # Get X and y
-
-    # %%
-    X = data.drop('count', axis=1)
-    y = data['count'].values
-
-
-    # %% [markdown]
-    # # First thing first: Spatio-temporal train test split
-
-    # %%
-    print('ST_train_test_split ...')
-    from stemflow.model_selection import ST_train_test_split
-    X_train, X_test, y_train, y_test = ST_train_test_split(X, y, 
-                                                        Spatio_blocks_count = 10, Temporal_blocks_count=10,
-                                                        random_state=42, test_size=0.3)
-    print('Done.')
-
-    # %% [markdown]
-    # # Train AdaSTEM hurdle model
-
-    # %%
-    print('Importing stemflow modules...')
-    from stemflow.model.AdaSTEM import AdaSTEM, AdaSTEMClassifier, AdaSTEMRegressor
-    from xgboost import XGBClassifier, XGBRegressor
-    from stemflow.model.Hurdle import Hurdle_for_AdaSTEM, Hurdle
-    print('Done.')
-
-    # %%
-    print('Declaring model instance...')
-    model = Hurdle_for_AdaSTEM(
-        classifier=AdaSTEMClassifier(base_model=XGBClassifier(tree_method='hist',random_state=42, verbosity = 0, n_jobs=1),
-                                    save_gridding_plot = True,
-                                    ensemble_fold=5, 
[…]
