+++ .github/workflows/compile_docs.yml
@@ -1,20 +1,23 @@
-name: ci 
+name: compile_docs 
+
 on:
   push:
+    paths:
+      - 'docs/**'
     branches:
-      - master 
       - main
+
 permissions:
   contents: write
 jobs:
-  deploy:
+  deploy-docs:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
         with:
           python-version: 3.x
-      - uses: actions/cache@v2
+      - uses: actions/cache@v3
         with:
           key: ${{ github.ref }}
           path: .cache
