+++ tests/test_all.py
@@ -1,0 +1,378 @@
+# %% [markdown]
+# # A **mini** test of stemflow
+# 
+# Yangkang Chen<br>
+# Sep 12, 2023
+
+# %%
+import pandas as pd
+import numpy as np
+from tqdm.auto import tqdm
+import matplotlib.pyplot as plt
+import pickle
+import os
+from urllib.request import urlopen
+import time
+
+# from stemflow.mini_test import run_mini_test
+
+# run_mini_test(delet_tmp_files=False, speed_up_times=2)
+
+def test_mini(delet_tmp_files: bool=True, show: bool = False, ensemble_models_disk_saver=False, ensemble_models_disk_saving_dir='./', speed_up_times=1):
+    """Run a mini test
+    
+    Processes:
+        1. Request data
+        2. ST_train_test_split
+        3. Import stemflow modules
+        4. Declare model instance
+        5. Fitting model
+        6. Calculating feature importances
+        7. Assigning importance to points
+        8. Plotting top 2 important variables
+        9. Calculate the fitting errors
+        10. Predicting on test set
+        11. Evaluation
+        12. Watermark
+        13. Deleting tmp files (optional)
+        
+    Args:
+        delet_tmp_files:
+            Whether to delet files after mini test.
+        show:
+            Whether to show the charts in jupyter.
+        ensemble_models_disk_saver:
+            Whether to save ensembles in disk instead of memory.
+        ensemble_models_disk_saving_dir:
+            if ensemble_models_disk_saver == True, where to save the models.
+        speed_up_times:
+            Speed up the mini test. For example, 2 means 2 times speed up. 
+    """
+    #
+    print('Start Running Mini-test...')
+    print(f'Temporary files will be stored at: ./stemflow_mini_test/')
+    if delet_tmp_files:
+        print('Temporary files will be deleted.')
+    else:
+        print('Temporary files will *NOT* be deleted.')
+    # download mini data
+    if not os.path.exists('./stemflow_mini_test'):
+        os.makedirs('./stemflow_mini_test')
+    if not 'mini_data.pkl' in os.listdir('./stemflow_mini_test'):
+        url = "https://chenyangkang.github.io/stemflow/mini_data/mini_data.pkl"
+        print(f'Requesting data from {url} ...')
+        data = pickle.load(urlopen(url))
+        # data = pd.read_csv(url)
+        data.to_csv('./stemflow_mini_test/mini_data.csv', index=False)
+        print('Done.')
+    else:
+        print('Mini-data already downloaded.')
+    
+    assert os.path.exists('./stemflow_mini_test/mini_data.csv')
+
+
+    # %%
+    # load data
+    data = pd.read_csv('./stemflow_mini_test/mini_data.csv')
+
+    # %%
+    # data.head()
+
+    # %%
+    plt.scatter(
+        data.longitude,
+        data.latitude,
+        s=0.2
+    )
+    plt.savefig('./stemflow_mini_test/data_plot.pdf')
+    
+    if show:
+        plt.show()
+    else:
+        plt.close()
+       
+    time.sleep(1)
+    assert os.path.exists('./stemflow_mini_test/data_plot.pdf')
+
+    # %% [markdown]
+    # # Get X and y
+
+    # %%
+    
+    x_names = [ 'duration_minutes',
+                'Traveling',
+                'DOY',
+                'time_observation_started_minute_of_day',
+                'elevation_mean',
+                'slope_mean',
+                'eastness_mean',
+                'northness_mean',
+                'bio1',
+                'bio2',
+                'bio3',
+                'bio4',
+                'bio5',
+                'bio6',
+                'bio7',
[…]
