+++ .github/workflows/run_pytest.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.7", "3.8", "3.9", "3.10"]
+        python-version: ["3.8", "3.9", "3.10"]
     steps:
     - uses: actions/checkout@v4
     - name: Set up Python ${{ matrix.python-version }}
@@ -31,14 +31,14 @@ jobs:
       run: |
         pytest -n auto --cov --no-cov-on-fail --cov-report=term-missing:skip-covered --cov-report xml:coverage.xml
 
-    - uses: action-badges/create-orphan-branch@0.1.0
-      with:
-        branch-name: badges
+    # - uses: action-badges/create-orphan-branch@0.1.0
+    #   with:
+    #     branch-name: badges
 
-    - name: Make Coverage Badge
-      uses: action-badges/cobertura-coverage-xml-badges@0.3.0
-      with:
-        file-name: coverage.svg
-        badge-branch: badges
-        github-token: '${{ secrets.SECRET_TOKEN }}'
-        coverage-file-name: ./coverage.xml
+    # - name: Make Coverage Badge
+    #   uses: action-badges/cobertura-coverage-xml-badges@0.3.0
+    #   with:
+    #     file-name: coverage.svg
+    #     badge-branch: badges
+    #     github-token: '${{ secrets.SECRET_TOKEN }}'
+    #     coverage-file-name: ./coverage.xml
