+++ MANIFEST.in
@@ -1,0 +1,7 @@
+include LICENSE
+include README.md
+include tests
+include requirements.txt
+include setup.py
+recursive-include stemflow *
+exclude docs/**
