+++ paper/md_paper/paper.bib
@@ -422,7 +422,3 @@
 	pages = {e05877},
 	file = {A. Lee‐Yaw et al. - 2022 - Species distribution models rarely predict the bio.pdf:/Users/chenyangkang/Zotero/storage/47MPKJDC/A. Lee‐Yaw et al. - 2022 - Species distribution models rarely predict the bio.pdf:application/pdf},
 }
-
-
-
-
