+++ stemflow/model/__init__.py
@@ -1,6 +1,6 @@
 from .static_func_AdaSTEM import (
     _monkey_patched_predict_proba,
-    train_one_stixel,
     assign_points_to_one_ensemble,
-    transform_pred_set_to_STEM_quad
+    train_one_stixel,
+    transform_pred_set_to_STEM_quad,
 )
