+++ tests/test_all.py
@@ -1,36 +1,46 @@
 # %% [markdown]
 # # A **mini** test of stemflow
-# 
+#
 # Yangkang Chen<br>
 # Sep 12, 2023
 
+import os
+import pickle
+import time
+from urllib.request import urlopen
+
+import matplotlib.pyplot as plt
+import numpy as np
+
 # %%
 import pandas as pd
-import numpy as np
 from tqdm.auto import tqdm
-import matplotlib.pyplot as plt
-import pickle
-import os
-from urllib.request import urlopen
-import time
 
 # from stemflow.mini_test import run_mini_test
 
 # run_mini_test(delet_tmp_files=False, speed_up_times=2)
 
-def load_data_from_doc():
-    if not os.path.exists('./tests/stemflow_mini_test'):
-        os.makedirs('./tests/stemflow_mini_test')
-        
-    print(os.getcwd())
-    cp_results = os.popen('cp ./docs/mini_data/mini_data.pkl ./tests/stemflow_mini_test', 'w')
-    time.sleep(1)
-    print(os.listdir('./tests/stemflow_mini_test'))
-    assert os.path.exists('./tests/stemflow_mini_test/mini_data.pkl')
 
-def test_mini(delet_tmp_files: bool=True, show: bool = False, ensemble_models_disk_saver=False, ensemble_models_disk_saving_dir='./', speed_up_times=1):
+def load_data_from_doc():
+    if not os.path.exists("./tests/stemflow_mini_test"):
+        os.makedirs("./tests/stemflow_mini_test")
+
+    print(os.getcwd())
+    _ = os.popen("cp ./docs/mini_data/mini_data.pkl ./tests/stemflow_mini_test", "w")
+    time.sleep(1)
+    print(os.listdir("./tests/stemflow_mini_test"))
+    assert os.path.exists("./tests/stemflow_mini_test/mini_data.pkl")
+
+
+def test_mini(
+    delet_tmp_files: bool = True,
+    show: bool = False,
+    ensemble_models_disk_saver=False,
+    ensemble_models_disk_saving_dir="./",
+    speed_up_times=1,
+):
     """Run a mini test
-    
+
     Processes:
         1. Request data
         2. ST_train_test_split
@@ -45,7 +55,7 @@ def test_mini(delet_tmp_files: bool=True, show: bool = False, ensemble_models_di
         11. Evaluation
         12. Watermark
         13. Deleting tmp files (optional)
-        
+
     Args:
         delet_tmp_files:
             Whether to delet files after mini test.
@@ -56,209 +66,216 @@ def test_mini(delet_tmp_files: bool=True, show: bool = False, ensemble_models_di
         ensemble_models_disk_saving_dir:
             if ensemble_models_disk_saver == True, where to save the models.
         speed_up_times:
-            Speed up the mini test. For example, 2 means 2 times speed up. 
+            Speed up the mini test. For example, 2 means 2 times speed up.
     """
     #
-    print('Start Running Mini-test...')
-    print(f'Temporary files will be stored at: ./tests/stemflow_mini_test/')
+    print("Start Running Mini-test...")
+    print("Temporary files will be stored at: ./tests/stemflow_mini_test/")
     if delet_tmp_files:
-        print('Temporary files will be deleted.')
+        print("Temporary files will be deleted.")
     else:
-        print('Temporary files will *NOT* be deleted.')
-        
+        print("Temporary files will *NOT* be deleted.")
+
     # Cp mini data
     load_data_from_doc()
-    
+
     # download mini data
-    if not os.path.exists('./tests/stemflow_mini_test'):
-        os.makedirs('./tests/stemflow_mini_test')
-    
[…]
