+++ stemflow/utils/generate_soft_colors.py
@@ -1,6 +1,8 @@
-import matplotlib
 import random
+
+import matplotlib
 import numpy as np
+
 
 def generate_soft_color() -> np.ndarray:
     """generating random soft colors
@@ -8,7 +10,7 @@ def generate_soft_color() -> np.ndarray:
     Returns:
         rgb (tuple): a rgb color
     """
-    
+
     # Generate random hue (between 0 and 1)
     hue = random.random()
 
@@ -17,7 +19,7 @@ def generate_soft_color() -> np.ndarray:
 
     # Generate a lower brightness value for a softer color
     brightness = random.uniform(0.6, 0.9)
-    
+
     rgb = matplotlib.colors.hsv_to_rgb([hue, saturation, brightness])
-    
+
     return rgb
