+++ requirements.txt
@@ -4,9 +4,9 @@ matplotlib>=3.7.1
 numpy>=1.24.3
 pandas>=1.5.3
 scikit_learn>=1.2.2
-# seaborn>=0.11.2
-setuptools>=68.2.0
-tqdm>=4.65.0
 # watermark>=2.3.1
 # xgboost>=1.7.6
 scipy>=1.10.1
+# seaborn>=0.11.2
+setuptools>=68.2.0
+tqdm>=4.65.0
