+++ docs/Fun_Visualization/Global_Mean_Temperature_2020.md
@@ -6,7 +6,3 @@ Notice: the visualized data is not modeled by AdaSTEM. Only visualization functi
 
 
 ![](./Temp_2020.gif)
-
-
-
-
