+++ .github/workflows/publish-pypi.yml
@@ -1,0 +1,39 @@
+# name: publish-pypi
+# on:
+#   push:
+#     paths:
+#       - 'stemflow/version.py'
+#     tags:
+#       - "v*.*.*.*"
+
+# jobs:
+#   publish:
+#     runs-on: ubuntu-latest
+#     environment:
+#       name: pypi
+#       url: https://pypi.org/p/stemflow/
+#     permissions:
+#       id-token: write
+#     steps:
+#       - uses: actions/checkout@v4
+#       - name: Python setup
+#         uses: actions/setup-python@v5
+#         with:
+#           python-version: 3.9
+#           cache: 'pip'
+#       - name: Install package
+#         run: |
+#           python -m pip install --upgrade pip
+#           pip install wheel
+#           pip install .
+#       - name: Build package
+#         run: |
+#           python setup.py sdist bdist_wheel
+#       - name: Publish to PyPI
+#         uses: pypa/gh-action-pypi-publish@release/v1
+#         with:
+#           user: __token__
+#           password: ${{ secrets.PYPI_API_TOKEN }}
+#           packages_dir: ./dist
+#           # repository-url: https://pypi.org/p/stemflow/
+#           verbose: true
