+++ mkdocs.yml
@@ -81,6 +81,9 @@ nav:
     - Examples/05.Hurdle_in_ada_or_ada_in_hurdle.ipynb
     - Examples/06.Base_model_choices.ipynb
     - Examples/07.Optimizing_Stixel_Size.ipynb
+  - Tips:
+    - 'Tips for data types': Tips/Tips_for_data_types.md
+    - 'Tips for different tasks': Tips/Tips_for_different_tasks.md
   - API Documentation:
     - stemflow.model:
       - 'AdaSTEM': API_Documentation/stemflow.model.AdaSTEM.md
