+++ docs/Input_data_structure_for_stemflow.md
@@ -1,3 +1,0 @@
-# Input data structure for stemflow
-
-## 
