+++ docs/Tips/Tips_for_data_types.md
@@ -4,6 +4,7 @@ In the both the [mini test](https://chenyangkang.github.io/stemflow/Examples/00.
 
 Here, we present more tips and examples on how to play with these indexing systems.
 
+------
 ## Flexible coordinate systems
 
 `stemflow` support all types of spatial coordinate reference system (CRS) and temporal indexing (for example, week month, year, or decades). `stemflow` only support tabular point data currently. You should transform your data to desired CRS before feeding them to `stemflow`.
@@ -70,6 +71,7 @@ pred = np.where(pred<0, 0, pred)
 eval_metrics = AdaSTEM.eval_STEM_res('classification',y_test, pred_mean)
 ```
 
+------
 ## Spatial-only modeling
 
 By playing some tricks, you can also do a `spatial-only` modeling, without splitting the data into temporal blocks:
@@ -86,8 +88,8 @@ model = AdaSTEMClassifier(
     grid_len_lat_lower_threshold=1e3,
     temporal_start=1,
     temporal_end=52,                            
-    temporal_step=2,
-    temporal_bin_interval=4,
+    temporal_step=1000,                 # Setting step and interval largely outweigh 
+    temporal_bin_interval=1000,         # temporal scale of data
     points_lower_threshold=50,             
     Spatio1='proj_lng',                   
     Spatio2='proj_lat',
@@ -97,10 +99,67 @@ model = AdaSTEMClassifier(
 )
 ```
 
+Setting `temporal_step` and `temporal_bin_interval` largely outweigh the temporal scale (1000 compared with 52) of your data will render only `one` temporal window during splitting. Consequently, your model would become a spatial model. This could be beneficial if temporal heterogeneity is not of interest, or without enough data to investigate.
 
-
-
-
+------
 ## Continuous and categorical features
 
+Basically, `stemflow` is a framework for spatial temporal indexing during modeling. It serves as a container to help `base model` do better jobs, and prevent distant modeling/prediction problem in space and time. Therefore, any feature you use during common tabular data modeling could be used here. It means that both continuous and categorical features can be the input, based on your expectation in the feature engineering.
+
+For categorical features, we recommend [one-hot encoding](https://en.wikipedia.org/wiki/One-hot) if the size of the category is not too large.
+
+Tree-based models (e.g., decision tree, boosting tree, random forest) are robust to missing values so you can fill the missing values with artificial values like `-1`. For other methods, there are different ways to fill the missing values, with [pros and cons](https://towardsdatascience.com/7-ways-to-handle-missing-values-in-machine-learning-1a6326adf79e).
+
+------
 ## Static and dynamic features
+
+### Concepts and examples
+Static features are summaries of status, while dynamic features vary largely across each almost each records.
+
+For example, for a task modeling the abundance of a bird each day in 2020;
+
+Static features:
+
+    - Land cover of year 2020:
+        - percentage cover of forest
+        - patch density of urban area
+
+    - Climate of year 2020:
+        - BIO1: Annual Mean Temperature
+        - BIO2: Mean Diurnal Range (Mean of monthly (max temp - min temp))
+        - BIO19: Precipitation of Coldest Quarter
+
+    - Normalized difference vegetation index (NDVI):
+        - NDVI_max: Highest NDVI of the annual cycle
+        - NDVI_std: Variation of NDVI in the annual cycle
+    
+
+Dynamic features:
+
+    - Weather of each checklist (record):
+        - the temperature of the hour (that we observed this bird)
+        - the total precipitation of the hour (that we observed this bird)
+        - V component of wind of the hour (that we observed this bird)
+
+    - Normalized difference vegetation index (NDVI):
+        - The absolute NDVI of the day (that we observed this bird)
+
+
+### Use of static and dynamic features
+
[…]
