### 2024-01-10 fc39759
add how to fix the gird size to tips and list stixel_training_size_threshold as an input variables of AdasTEM
