### 2024-01-10 0b06ca1
Working on Tips for different tasks
