### 2024-01-10 a806e1f
Working on Tips for different tasks
