+++ docs/Tips/Tips_for_different_tasks.md
@@ -1,9 +1,134 @@
 # Tips for different tasks
 
-## Regression and classification
+`stemflow` supports different types of tabular data modeling tasks, including
 
-TODO
+- Binary classification
+- Regression
+- Hurdle regression (first classify then regress on the positive part) for zero-inflated data
+
+
+## Classification and regression
+
+To create a classification or regression model, you can simply use the corresponding model classes.
+
+### Binary classification
+
+#### Create a model
+
+For binary classification, you can pass a sklearn `BaseEstimator` style classifier (in this case `XGBClassifier`) to the AdaSTEM classifier wrapper (`AdaSTEMClassifier`):
+
+```python
+from stemflow.model.AdaSTEM import AdaSTEM, AdaSTEMClassifier
+from xgboost import XGBClassifier
+
+model = AdaSTEMClassifier(
+    base_model=XGBClassifier(tree_method='hist',random_state=42, verbosity = 0,n_jobs=1),
+    save_gridding_plot = True,
+    ensemble_fold=10,
+    min_ensemble_required=7,
+    grid_len_lon_upper_threshold=1e5,
+    grid_len_lon_lower_threshold=1e3,
+    grid_len_lat_upper_threshold=1e5,
+    grid_len_lat_lower_threshold=1e3,
+    temporal_start=1,
+    temporal_end=366,                            
+    temporal_step=20,  
+    temporal_bin_interval=50, 
+    points_lower_threshold=50,             
+    Spatio1='proj_lng',                   
+    Spatio2='proj_lat',
+    Temporal1='DOY',
+    use_temporal_to_train=True,
+    njobs=1
+)
+```
+
+#### Prediction
+
+```py
+## fit
+model = model.fit(X_train.reset_index(drop=True), y_train)
+
+## predict
+pred = model.predict(X_test)
+```
+
+Alternatively, you can predict the probability:
+
+```py
+## get class probability
+pred_proba = model.predict_proba(X_test)
+```
+
+Or return the prediction variation across the ensembles:
+
+```py
+## also return the prediction std
+pred_proba_mean, pred_proba_std = model.predict_proba(X_test, return_std=True)
+```
+
+
+#### Evaluation
+
+Correspondingly, you would use a set of metrics for the classification problem:
+```
+# Evaluation
+eval_metrics = AdaSTEM.eval_STEM_res('classification',y_test, pred_mean)
+```
+This will return a bunch of metrics, including AUC, precision, recall, etc.
+
+
+
+### Regression
+
+For Regression problem, you can pass a sklearn `BaseEstimator` style regressor (in this case `XGBRegressor`) to the AdaSTEM regressor wrapper (`AdaSTEMRegressor`):
+
+```python
+from stemflow.model.AdaSTEM import AdaSTEM, AdaSTEMRegressor
+from xgboost import XGBRegressor
+
+model = AdaSTEMRegressor(
+    base_model=XGBRegressor(tree_method='hist',random_state=42, verbosity = 0,n_jobs=1),
+    save_gridding_plot = True,
+    ensemble_fold=10,
+    min_ensemble_required=7,
+    grid_len_lon_upper_threshold=1e5,
+    grid_len_lon_lower_threshold=1e3,
+    grid_len_lat_upper_threshold=1e5,
+    grid_len_lat_lower_threshold=1e3,
+    temporal_start=1,
+    temporal_end=366,                            
+    temporal_step=20,  
+    temporal_bin_interval=50, 
+    points_lower_threshold=50,             
+    Spatio1='proj_lng',                   
+    Spatio2='proj_lat',
+    Temporal1='DOY',
+    use_temporal_to_train=True,
+    njobs=1
+)
+```
[…]
