### 2024-01-10 eaf5434
complete Tips for different tasks
