+++ docs/Tips/Tips_for_different_tasks.md
@@ -43,6 +43,8 @@ model = AdaSTEMClassifier(
 )
 ```
 
+`stemflow` automatically calculates binary class weights (using `sklearn.utils.class_weight.compute_class_weight` with `class_weight` set as `'balanced'`) in each stixels for imbalanced data and pass them to base model during training.
+
 #### Prediction
 
 ```py
@@ -129,6 +131,53 @@ Likewise, you could also return the variation of prediction by setting `return_s
 
 [Hurdle model](https://en.wikipedia.org/wiki/Hurdle_model#:~:text=A%20hurdle%20model%20is%20a,of%20the%20non%2Dzero%20values.) is different from regression or classification model – it combines them.
 
-Hurdle model is designed to solve the [zero-inflation problems](https://en.wikipedia.org/wiki/Zero-inflated_model), which is commonly seen in count data.
+Hurdle model is designed to solve the [zero-inflation problems](https://en.wikipedia.org/wiki/Zero-inflated_model), which is commonly seen in **count data**, for example, the count of individual birds for a species in each checklists. Too many "zeros" in target data will bias the loss function and induce imbalanced likelihood, weaken the model performance. Similar methods to solve zero-inflation problems include Zero-Inflated Poisson Model (ZIP), Zero-Inflated Negative Binomial Model (ZINB).
 
-Similar methods to solve zero-inflation problems include Zero-Inflated Poisson Model (ZIP), Zero-Inflated Negative Binomial Model (ZINB).
+### Hurdle model workflow
+Hurdle model, as its name indicates, uses two-step modeling method to solve the zero-inflation problems:
+
+![Hurdle model workflow](../Hurdle_workflow.png)
+
+As shown in this figure, zero-inflated input data was first used to train a classifier, with labels being binary transformed. Then, the samples with positive labels are used to train a regressor. Together these two compose a hurdle model. When new data comes in, classifier and regressor make prediction independently, and the results are combined/masked to yield the final prediction (only those samples classified as positive will be assigned continuous prediction from the regressor).
+
+### Hurdle model common practice
+Combined with AdaSTEM framework, hurdle model is usually used as a based model:
+
+```python
+model = AdaSTEMRegressor(
+    base_model=Hurdle(
+        classifier=XGBClassifier(...),
+        regressor=XGBRegressor(...)
+    ),  
+    ...
+)
+```
+
+This is a common practice in most paper I've seen, including reference [1-3].
+
+However, AdaSTEM model could also be used as regressor or classifier is needed:
+
+```python
+model_Ada_in_Hurdle = Hurdle_for_AdaSTEM(
+    classifier=AdaSTEMClassifier(
+        base_model=XGBClassifier(...),
+        ...),
+    regressor=AdaSTEMRegressor(
+        base_model=XGBRegressor(...),
+        ...)
+)
+
+```
+
+The later one is more a conservative framework and have higher `precision` score, but lower recall and overall performance. For more details see "[Hurdle in AdaSTEM or AdaSTEM in hurdle?](https://chenyangkang.github.io/stemflow/Examples/05.Hurdle_in_ada_or_ada_in_hurdle.html)"
+
+
+
+-----
+References:
+
+1. [Fink, D., Damoulas, T., & Dave, J. (2013, June). Adaptive Spatio-Temporal Exploratory Models: Hemisphere-wide species distributions from massively crowdsourced eBird data. In Proceedings of the AAAI Conference on Artificial Intelligence (Vol. 27, No. 1, pp. 1284-1290).](https://ojs.aaai.org/index.php/AAAI/article/view/8484)
+
[…]
