### 2024-01-10 0f83154
complete Tips for different tasks
