### 2024-01-10 44af95b
v1.0.9.3; Fix quadtree rotation angle problem
