+++ .pre-commit-config.yaml
@@ -26,7 +26,7 @@ repos:
       - id: check-yaml
         exclude: recipe/meta.yaml
       - id: check-added-large-files
-        args: ["--maxkb=1024"]
+        args: ["--maxkb=10240"]
       - id: detect-private-key
       - id: end-of-file-fixer
       - id: requirements-txt-fixer
