### 2024-01-11 8d9268a
fix AdaSTEM temporal window overlapping probelm
