### 2024-01-12 0a10264
Add annotation to 01.AdaSTEM_demo.ipynb
