### 2024-01-12 4291ae0
Add annotation to 01.AdaSTEM_demo.ipynb
