+++ docs/index.md
@@ -184,7 +184,7 @@ Daily Abundance Map of Barn Swallow
 
 ![GIF visualization](https://github.com/chenyangkang/stemflow/raw/main/docs/pred_gif.gif)
 
-See section [Prediction and Visualization](https://chenyangkang.github.io/stemflow/Examples/04.Prediction_visualization.html) for how to generate this GIF.
+See section [AdaSTEM demo](https://chenyangkang.github.io/stemflow/Examples/01.AdaSTEM_demo.html) for how to generate this GIF.
 
 ----
 
