### 2024-01-13 chenyangkang/stemflow#2 @chenyangkang
Hi Chris,

Thank you for your valuable comments. I agree on most of the issues you mentioned. Here are the replies and solutions for them:

1. About how time and space are encoded:

> Do input dataframes require a DOY column?

No, the target spatial and temporal indexing columns are flexible, and could be passed as arguments.

> Does anything change if data are provided at other temporal scales (weekly, monthly, yearly, etc.)?

No, the scale of the indexing is highly flexible and user-defined.

> I see latitude and longitude encoded in the mini test data - is that the only supported CRS?

No. Any indexing system could be passed to AdaSTEM framework. If user wish to use other CRS, they could transform it in prior.

 > I see that geopandas is a dependency - does the model class support passing a GeoDataFrame, or do you need to explicitly encode column names?

This is a mistake. `geopandas` is not a dependency of `stemflow`. I removed them from requirement, for both pypi and conda. Currently we only support `pandas` dataframe as training data input (there is a type check). Geopandas dataframe can be tranfomed to `pd.core.frame.DataFrame` beforehand.

PR #14 was to add additional documentations on these issues.

All these issues are documented in the new documentation page [Tips for data types](https://chenyangkang.github.io/stemflow/Tips/Tips_for_data_types.html) as well as listed as features in the [home page](https://chenyangkang.github.io/stemflow/) (section `Model and data`).
