### 2024-01-13 chenyangkang/stemflow#2 @chenyangkang
2. Thank you for your advice on **more guidance for feature data**. Here comes the replies and solutions:

> But it looks like the mini_data example is nearly all static features, with DOY being the only dynamic variable. Is this how other datasets are expected to be formatted? Can users provide a combination of static and dynamic covariates?

This is a great question. You are right, the datasets are formatted with static features except for `DOY`. This is for specific reasons. Including:

- Model assumption
- Scale of interest
- Consideration for overfitting
- Align with prediction set

However, dynamic features are absolutely suitable for training, or combination of static and dynamic ones.

The details are discussed in the additional documentation [Tips for data types](https://chenyangkang.github.io/stemflow/Tips/Tips_for_data_types.html).

> Are categorical features supported? More clarification regarding best practices for how to extract and format covariate data would help a lot.

Yes, categorical features are supported. We mentioned the feature formatting in [Tips for data types](https://chenyangkang.github.io/stemflow/Tips/Tips_for_data_types.html).
