### 2024-01-13 1b5a424
add base model choies tips
