### 2024-01-13 chenyangkang/stemflow#2 @chenyangkang
4. Thank you for your comments on `adding more explanation about functionality`

> I'm not familiar with the Hurdle modeling approach, which seems great and appears foundational to AdaSTEM - tips on best practices here would be valuable.

Definitely true. I add more information about what tasks `stemflow` could do, especially what is `hurdle` in the new document [Tips for different tasks](https://chenyangkang.github.io/stemflow/Tips/Tips_for_different_tasks.html). We also mention the function repertoire at the [home page](https://chenyangkang.github.io/stemflow/index.html).

> You mentioned using other base models from sklearn in the manuscript - when would this be advantageous?

Thank you for your questions. We add additional analysis and bump chart for comparing the performance of using different base model in [Base model choices](https://chenyangkang.github.io/stemflow/Examples/06.Base_model_choices.html). Basically the trade-off exists among "performance – speed – memory usage". `stemflow` is a bagging modeling framework, which makes it time-consuming and memory-consuming. We would want the best trade-off between performance and resources consumption.

> You also provide great examples for comparing learning curves and optimizing strixel sizes, but these notebooks are just code blocks with some plots and little interpretation. I found myself wanting more guidance for how to interpret these results, or even guidance on why such optimization is important (do you optimize strixel size to minimize overfitting, for example?).

Thank you for your suggestions. I add additional charts, comments, and conclusion at section [Optimizing stixel size](https://chenyangkang.github.io/stemflow/Examples/07.Optimizing_stixel_size.html) for addressing `why such optimization is important` and `how to optimize the stixels`. I hope this helps users in understanding the parameter choices.
