### 2024-01-13 chenyangkang/stemflow#3 @chenyangkang
Thank you for the suggestion! I agree that the test should be automatic.

I have made the mini_test [automatic at each PR and push](https://github.com/chenyangkang/stemflow/blob/main/.github/workflows/run_pytest.yml), testing on Python 3.8, 3.9, and 3.10.

I still kept the `stemflow.manually_testing` module and [mini_test notebook](https://chenyangkang.github.io/stemflow/Examples/00.Mini_test.html) in case any user is interested in it.

Besides, I add pre-commit for code trimming and auto-checking, as well as codecov statistics.

Related PRs: #9, #11 , #12
