### 2024-01-13 chenyangkang/stemflow#4 @chenyangkang
Thank you for your comments. I added contributing documentation [here](https://github.com/chenyangkang/stemflow/blob/main/docs/CONTRIBUTING.md), as well as [code of conduct](https://github.com/chenyangkang/stemflow/blob/main/docs/CODE_OF_CONDUCT.md) and issue templates.

Related PRs: #5, #7, #8, #13
