### 2024-01-17 chenyangkang/stemflow#22 @chenyangkang
[JOSS review] paper revision
This issue is opened for paper revision suggested by [at]jedalong (https://github.com/openjournals/joss-reviews/issues/6158)

original comments:

> L10 - I think you need to explain what AdaSTEM is more explicitly at the beginning of the abstract similar to line 32 so the reader knows what this is.
> L18 consider rephrasing this sentence.
> L26 - consider rephrasing 'mine its merits'
> L79 - rephrase 'mounting'
> L85 - rephrase this entire sentence, not sure this is what you mean, confusing 'dependency conjugated with bias in data abundance'?
> L89 - rephrase 'potentials'
