### 2024-01-17 chenyangkang/stemflow#24 @chenyangkang
[JOSS review] Add spherical indexing system
This issue is to solve the problem suggested by [at]jedalong (https://github.com/openjournals/joss-reviews/issues/6158)

> I think it is probably associated with the tools used for modelling that do not expect spatial objects, but binning the data correctly based on spatial objects and then back transforming to the expected data structure migth be a more correct way to deal with the spherical geometry issue.

TODO:
1. Implement spherical indexing system
