### 2024-01-17 chenyangkang/stemflow#25 @chenyangkang
[JOSS review] Documentation revision
This issue is to solve the problem suggested by [at]jedalong (https://github.com/openjournals/joss-reviews/issues/6158)

> What is the difference between temporal_step and temporal_bin_interval? I didnt follow the whole 'sliding' window part of this are they bins/blocks or a moving window?
>
> Documentation Notes:
> Rephrase: stemflow have 4 important gridding parameters. Actually only two:
> The maximum grid length, and the minimum grid length. It can be separately set by longitude and latitude, and that will be 4.

TODO:
1. Add documentation on sliding window.
2. Add docs on the difference between `temporal_step` and `temporal_bin_interval `.
3. Revise the gridding params docs.
