### 2024-01-17 c905532
fix #25 and partial to #23
