### 2024-01-17 036409b
fix #25 and partial to #23
