+++ stemflow/model/STEM.py
@@ -1,0 +1,1296 @@
+import os
+import pickle
+import warnings
+from itertools import repeat
+
+#
+from multiprocessing import Pool, cpu_count
+from typing import Tuple, Union
+
+import numpy as np
+import pandas as pd
+from numpy import ndarray
+
+# validation check
+from pandas.core.frame import DataFrame
+from scipy.stats import pearsonr, spearmanr
+from sklearn.base import BaseEstimator
+from sklearn.metrics import (
+    average_precision_score,
+    cohen_kappa_score,
+    d2_tweedie_score,
+    f1_score,
+    mean_absolute_error,
+    mean_squared_error,
+    precision_score,
+    r2_score,
+    recall_score,
+    roc_auc_score,
+)
+from tqdm import tqdm
+
+#
+from ..utils.quadtree import get_ensemble_quadtree
+
+#
+from .AdaSTEM import AdaSTEM
+from .dummy_model import dummy_model1
+
+# from ..utils.validation import check_random_state
+from .static_func_AdaSTEM import (  # predict_one_ensemble
+    _monkey_patched_predict_proba,
+    assign_points_to_one_ensemble,
+    get_model_and_stixel_specific_x_names,
+    predict_one_stixel,
+    train_one_stixel,
+    transform_pred_set_to_STEM_quad,
+)
+
+#
+
+
+class STEM(AdaSTEM):
+    """A STEM model class inherited by STEMClassifier and STEMRegressor"""
+
+    def __init__(
+        self,
+        base_model: BaseEstimator,
+        task: str = "hurdle",
+        ensemble_fold: int = 10,
+        min_ensemble_required: int = 7,
+        grid_len: Union[float, int] = 25,
+        points_lower_threshold: int = 50,
+        stixel_training_size_threshold: int = None,
+        temporal_start: Union[float, int] = 1,
+        temporal_end: Union[float, int] = 366,
+        temporal_step: Union[float, int] = 20,
+        temporal_bin_interval: Union[float, int] = 50,
+        temporal_bin_start_jitter: Union[float, int, str] = "random",
+        spatio_bin_jitter_magnitude: Union[float, int] = 100,
+        save_gridding_plot: bool = True,
+        save_tmp: bool = False,
+        save_dir: str = "./",
+        sample_weights_for_classifier: bool = True,
+        Spatio1: str = "longitude",
+        Spatio2: str = "latitude",
+        Temporal1: str = "DOY",
+        use_temporal_to_train: bool = True,
+        njobs: int = 1,
+        subset_x_names: bool = False,
+        ensemble_models_disk_saver: bool = False,
+        ensemble_models_disk_saving_dir: str = "./",
+        plot_xlims: Tuple[Union[float, int], Union[float, int]] = (-180, 180),
+        plot_ylims: Tuple[Union[float, int], Union[float, int]] = (-90, 90),
+        verbosity: int = 0,
+    ):
+        """Make an AdaSTEM object
+
+        Args:
+            base_model:
+                base model estimator
+            task:
+                task of the model. One of 'classifier', 'regressor' and 'hurdle'. Defaults to 'hurdle'.
+            ensemble_fold:
+                Ensembles count. Higher, better for the model performance. Time complexity O(N). Defaults to 10.
+            min_ensemble_required:
+                Only points with more than this number of model ensembles available are predicted.
+                In the training phase, if stixels contain less than `points_lower_threshold` of data records,
+                the results are set to np.nan, making them `unpredictable`. Defaults to 7.
+            grid_len:
+                length of the grids. Defaults to 25.
+            points_lower_threshold:
+                Do not further split the gird if split results in less samples than this threshold.
+                Overrided by grid_len_*_upper_threshold parameters. Defaults to 50.
[…]
