+++ stemflow/utils/QTree.py
@@ -1,0 +1,396 @@
+import os
+import warnings
+from collections.abc import Sequence
+
+# from multiprocessing import Pool
+from typing import Tuple, Union
+
+import matplotlib.patches as patches
+import matplotlib.pyplot as plt  # plotting libraries
+import numpy as np
+import pandas
+import pandas as pd
+
+from .generate_soft_colors import generate_soft_color
+from .validation import check_random_state
+
+os.environ["MKL_NUM_THREADS"] = "1"
+os.environ["NUMEXPR_NUM_THREADS"] = "1"
+os.environ["OMP_NUM_THREADS"] = "1"
+
+warnings.filterwarnings("ignore")
+
+
+class Point:
+    """A Point class for recording data points"""
+
+    def __init__(self, index, x, y):
+        self.x = x
+        self.y = y
+        self.index = index
+
+
+class Node:
+    """A tree-like division node class"""
+
+    def __init__(
+        self, x0: Union[float, int], y0: Union[float, int], w: Union[float, int], h: Union[float, int], points: Point
+    ):
+        self.x0 = x0
+        self.y0 = y0
+        self.width = w
+        self.height = h
+        self.points = points
+        self.children = []
+
+    def get_width(self):
+        return self.width
+
+    def get_height(self):
+        return self.height
+
+    def get_points(self):
+        return self.points
+
+
+def recursive_subdivide(
+    node: Node,
+    grid_len_lon_upper_threshold: Union[float, int],
+    grid_len_lon_lower_threshold: Union[float, int],
+    grid_len_lat_upper_threshold: Union[float, int],
+    grid_len_lat_lower_threshold: Union[float, int],
+    points_lower_threshold: Union[float, int],
+):
+    """recursively subdivide the grids
+
+    Args:
+        node:
+            node input
+        grid_len_lon_upper_threshold:
+            force divide if grid longitude larger than the threshold
+        grid_len_lon_lower_threshold:
+            stop divide if grid longitude **will** be below than the threshold
+        grid_len_lat_upper_threshold:
+            force divide if grid latitude larger than the threshold
+        grid_len_lat_lower_threshold:
+            stop divide if grid latitude **will** be below than the threshold
+
+    """
+
+    if (node.width / 2 < grid_len_lon_lower_threshold) or (node.height / 2 < grid_len_lat_lower_threshold):
+        return
+
+    w_ = float(node.width / 2)
+    h_ = float(node.height / 2)
+
+    p = contains(node.x0, node.y0, w_, h_, node.points)
+    x1 = Node(node.x0, node.y0, w_, h_, p)
+    recursive_subdivide(
+        x1,
+        grid_len_lon_upper_threshold,
+        grid_len_lon_lower_threshold,
+        grid_len_lat_upper_threshold,
+        grid_len_lat_lower_threshold,
+        points_lower_threshold,
+    )
+
+    p = contains(node.x0, node.y0 + h_, w_, h_, node.points)
+    x2 = Node(node.x0, node.y0 + h_, w_, h_, p)
+    recursive_subdivide(
+        x2,
+        grid_len_lon_upper_threshold,
+        grid_len_lon_lower_threshold,
+        grid_len_lat_upper_threshold,
+        grid_len_lat_lower_threshold,
+        points_lower_threshold,
+    )
+
+    p = contains(node.x0 + w_, node.y0, w_, h_, node.points)
+    x3 = Node(node.x0 + w_, node.y0, w_, h_, p)
+    recursive_subdivide(
+        x3,
+        grid_len_lon_upper_threshold,
+        grid_len_lon_lower_threshold,
[…]

[…]
