### 2024-01-17 b526594
update testing for the changed grid_len configuration
