+++ stemflow/manually_testing.py
@@ -191,10 +191,8 @@ def run_mini_test(
         save_gridding_plot=True,
         ensemble_fold=fold_,
         min_ensemble_required=min_req,
-        grid_len_lon_upper_threshold=40,
-        grid_len_lon_lower_threshold=5,
-        grid_len_lat_upper_threshold=40,
-        grid_len_lat_lower_threshold=5,
+        grid_len_upper_threshold=40,
+        grid_len_lower_threshold=5,
         temporal_start=1,
         temporal_end=366,
         temporal_step=30,
