### 2024-01-18 10cc151
Add STEM class and update mini_test & API docs & #27
