+++ stemflow/manually_testing.py
@@ -17,45 +17,7 @@ import pandas as pd
 from tqdm.auto import tqdm
 
 
-def run_mini_test(
-    delet_tmp_files: bool = True,
-    show: bool = False,
-    ensemble_models_disk_saver=False,
-    ensemble_models_disk_saving_dir="./",
-    speed_up_times=1,
-    tmp_dir="./stemflow_mini_test",
-):
-    """Run a mini test
-
-    Processes:
-        1. Request data
-        2. ST_train_test_split
-        3. Import stemflow modules
-        4. Declare model instance
-        5. Fitting model
-        6. Calculating feature importances
-        7. Assigning importance to points
-        8. Plotting top 2 important variables
-        9. Calculate the fitting errors
-        10. Predicting on test set
-        11. Evaluation
-        12. Watermark
-        13. Deleting tmp files (optional)
-
-    Args:
-        delet_tmp_files:
-            Whether to delet files after mini test.
-        show:
-            Whether to show the charts in jupyter.
-        ensemble_models_disk_saver:
-            Whether to save ensembles in disk instead of memory.
-        ensemble_models_disk_saving_dir:
-            if ensemble_models_disk_saver == True, where to save the models.
-        speed_up_times:
-            Speed up the mini test. For example, 2 means 2 times speed up.
-    """
-    #
-    print("Start Running Mini-test...")
+def get_data(delet_tmp_files, tmp_dir):
     print("Temporary files will be stored at: ./stemflow_mini_test/")
     if delet_tmp_files:
         print("Temporary files will be deleted.")
@@ -78,23 +40,10 @@ def run_mini_test(
 
     assert os.path.exists(len(data) > 0)
 
-    # %%
-    plt.scatter(data.longitude, data.latitude, s=0.2)
-    plt.savefig(os.path.join(tmp_dir, "data_plot.pdf"))
+    return data
 
-    if show:
-        plt.show()
-    else:
-        plt.close()
 
-    time.sleep(1)
-    assert os.path.exists(os.path.join(tmp_dir, "data_plot.pdf"))
-
-    # %% [markdown]
-    # # Get X and y
-
-    # %%
-
+def get_x_names():
     x_names = [
         "duration_minutes",
         "Traveling",
@@ -141,20 +90,17 @@ def run_mini_test(
         "woody_savannas",
         "entropy",
     ]
+    return x_names
 
-    X = data.drop("count", axis=1)[x_names + ["longitude", "latitude"]]
-    y = data["count"].values
 
-    # %% [markdown]
-    # # First thing first: Spatio-temporal train test split
-
-    # %%
+def get_path():
     import stemflow
 
     isl_path = stemflow.__path__[0]
     print(f"Installation path: {isl_path}")
 
-    #
+
+def ST_train_test_split(X, y):
     print("ST_train_test_split ...")
     from stemflow.model_selection import ST_train_test_split
 
@@ -163,25 +109,43 @@ def run_mini_test(
     )
     print("Done.")
     assert len(X_train) > 0 and len(X_test) > 0 and len(y_train) > 0 and len(y_test) > 0
+    return X_train, X_test, y_train, y_test
 
-    # %% [markdown]
-    # # Train AdaSTEM hurdle model
 
-    # %%
-    print("Importing stemflow modules...")
+def make_AdaSTEM_model1(fold_, min_req, ensemble_models_disk_saver, ensemble_models_disk_saving_dir):
+    from xgboost import XGBClassifier, XGBRegressor
+
+    from stemflow.model.STEM import STEM, STEMClassifier, STEMRegressor
+
+    model = STEMClassifier(
[…]
