+++ docs/API_Documentation/stemflow.model.STEM.md
@@ -1,0 +1,4 @@
+# stemflow.model.STEM
+---
+::: stemflow.model.STEM
+---
