### 2024-01-18 c3caa0b
Add STEM class and update mini_test & API docs & #27
