+++ stemflow/gridding/QuadGrid.py
@@ -1,0 +1,210 @@
+import os
+import warnings
+from collections.abc import Sequence
+
+# from multiprocessing import Pool
+from typing import Tuple, Union
+
+import matplotlib.patches as patches
+import matplotlib.pyplot as plt  # plotting libraries
+import numpy as np
+import pandas
+import pandas as pd
+
+from ..utils.generate_soft_colors import generate_soft_color
+from ..utils.validation import check_random_state
+from .Q_blocks import Grid, Node, Point
+
+os.environ["MKL_NUM_THREADS"] = "1"
+os.environ["NUMEXPR_NUM_THREADS"] = "1"
+os.environ["OMP_NUM_THREADS"] = "1"
+
+warnings.filterwarnings("ignore")
+
+
+class QuadGrid:
+    """A QuadTree class"""
+
+    def __init__(
+        self,
+        grid_len: Union[float, int],
+        points_lower_threshold: int,
+        lon_lat_equal_grid: bool = True,
+        rotation_angle: Union[float, int] = 0,
+        calibration_point_x_jitter: Union[float, int] = 0,
+        calibration_point_y_jitter: Union[float, int] = 0,
+    ):
+        self.points_lower_threshold = points_lower_threshold
+        self.grid_len = grid_len
+        self.lon_lat_equal_grid = lon_lat_equal_grid  # Not used anyway
+        self.points = []
+        self.rotation_angle = rotation_angle
+        self.calibration_point_x_jitter = calibration_point_x_jitter
+        self.calibration_point_y_jitter = calibration_point_y_jitter
+
+    def add_lon_lat_data(self, indexes: Sequence, x_array: Sequence, y_array: Sequence):
+        """Store input lng lat data and transform to **Point** object
+
+        Parameters:
+            indexes: Unique identifier for indexing the point.
+            x_array: longitudinal values.
+            y_array: latitudinal values.
+
+        """
+        if not len(x_array) == len(y_array) or not len(x_array) == len(indexes):
+            raise ValueError("input longitude and latitude and indexes not in same length!")
+
+        data = np.array([x_array, y_array]).T
+        angle = self.rotation_angle
+        r = angle / 360
+        theta = r * np.pi * 2
+        rotation_matrix = np.array([[np.cos(theta), -np.sin(theta)], [np.sin(theta), np.cos(theta)]])
+        data = data @ rotation_matrix
+        lon_new = (data[:, 0] + self.calibration_point_x_jitter).tolist()
+        lat_new = (data[:, 1] + self.calibration_point_y_jitter).tolist()
+
+        for index, lon, lat in zip(indexes, lon_new, lat_new):
+            self.points.append(Point(index, lon, lat))
+
+    def generate_gridding_params(self):
+        """For completeness"""
+        pass
+
+    def get_points(self):
+        """For completeness"""
+        pass
+
+    def subdivide(self):
+        """Called subdivide, but actually iterative divide"""
+
+        # Generate grids
+        x_list = np.array([i.x for i in self.points])
+        y_list = np.array([i.y for i in self.points])
+        xmin = np.min(x_list)
+        xmax = np.max(x_list)
+        ymin = np.min(y_list)
+        ymax = np.max(y_list)
+
+        self.x_start = xmin - self.grid_len + self.calibration_point_x_jitter
+        self.x_end = xmax + self.grid_len + self.calibration_point_x_jitter
+        self.y_start = ymin - self.grid_len + self.calibration_point_y_jitter
+        self.y_end = ymax + self.grid_len + self.calibration_point_y_jitter
+        x_grids = np.arange(self.x_start, self.x_end, self.grid_len)
+        y_grids = np.arange(self.y_start, self.y_end, self.grid_len)
+
[…]
