+++ stemflow/model/STEM.py
@@ -1,56 +1,23 @@
-import os
-import pickle
-import warnings
-from itertools import repeat
-
-#
-from multiprocessing import Pool, cpu_count
 from typing import Tuple, Union
 
-import numpy as np
-import pandas as pd
-from numpy import ndarray
-
-# validation check
-from pandas.core.frame import DataFrame
-from scipy.stats import pearsonr, spearmanr
 from sklearn.base import BaseEstimator
-from sklearn.metrics import (
-    average_precision_score,
-    cohen_kappa_score,
-    d2_tweedie_score,
-    f1_score,
-    mean_absolute_error,
-    mean_squared_error,
-    precision_score,
-    r2_score,
-    recall_score,
-    roc_auc_score,
-)
-from tqdm import tqdm
 
-#
-from ..utils.quadtree import get_ensemble_quadtree
-
-#
-from .AdaSTEM import AdaSTEM
-from .dummy_model import dummy_model1
-
-# from ..utils.validation import check_random_state
-from .static_func_AdaSTEM import (  # predict_one_ensemble
-    _monkey_patched_predict_proba,
-    assign_points_to_one_ensemble,
-    get_model_and_stixel_specific_x_names,
-    predict_one_stixel,
-    train_one_stixel,
-    transform_pred_set_to_STEM_quad,
-)
+from .AdaSTEM import AdaSTEM, AdaSTEMClassifier, AdaSTEMRegressor
 
 #
 
 
 class STEM(AdaSTEM):
-    """A STEM model class inherited by STEMClassifier and STEMRegressor"""
+    """A STEM model class (allow fixed grid size)
+
+    Parents:
+        stemflow.model.AdaSTEM
+
+    Children:
+        stemflow.model.STEM.STEMClassifier
+        stemflow.model.AdaSTEM.STEMRegressor
+
+    """
 
     def __init__(
         self,
@@ -58,7 +25,7 @@ class STEM(AdaSTEM):
         task: str = "hurdle",
         ensemble_fold: int = 10,
         min_ensemble_required: int = 7,
-        grid_len: Union[float, int] = 25,
+        grid_len: [float, int, None] = None,
         points_lower_threshold: int = 50,
         stixel_training_size_threshold: int = None,
         temporal_start: Union[float, int] = 1,
@@ -83,7 +50,7 @@ class STEM(AdaSTEM):
         plot_ylims: Tuple[Union[float, int], Union[float, int]] = (-90, 90),
         verbosity: int = 0,
     ):
-        """Make an AdaSTEM object
+        """Make a STEM object
 
         Args:
             base_model:
@@ -100,7 +67,7 @@ class STEM(AdaSTEM):
                 length of the grids. Defaults to 25.
             points_lower_threshold:
                 Do not further split the gird if split results in less samples than this threshold.
-                Overrided by grid_len_*_upper_threshold parameters. Defaults to 50.
+                Overriden by grid_len_*_upper_threshold parameters. Defaults to 50.
             stixel_training_size_threshold:
                 Do not train the model if the available data records for this stixel is less than this threshold,
                 and directly set the value to np.nan. Defaults to 50.
@@ -176,878 +143,64 @@ class STEM(AdaSTEM):
                 feature importance dataframe for each stixel.
 
         """
-        # save base model
-        self.base_model = base_model
-        self.Spatio1 = Spatio1
-        self.Spatio2 = Spatio2
-        self.Temporal1 = Temporal1
-        self.use_temporal_to_train = use_temporal_to_train
-        self.subset_x_names = subset_x_names
-        self.ensemble_models_disk_saver = ensemble_models_disk_saver
-        self.ensemble_models_disk_saving_dir = ensemble_models_disk_saving_dir
-        if self.ensemble_models_disk_saver:
[…]
