+++ docs/API_Documentation/stemflow.utils.quadtree.md
@@ -1,5 +1,5 @@
 # stemflow.utils.quadtree
 
 ---
-:::stemflow.utils.quadtree
+::: stemflow.utils.quadtree
 ---
