### 2024-01-18 f0852f8
fix docs for the STEM & AdaSTEM
