+++ docs/Tips/Tips_for_different_tasks.md
@@ -26,10 +26,8 @@ model = AdaSTEMClassifier(
     save_gridding_plot = True,
     ensemble_fold=10,
     min_ensemble_required=7,
-    grid_len_lon_upper_threshold=1e5,
-    grid_len_lon_lower_threshold=1e3,
-    grid_len_lat_upper_threshold=1e5,
-    grid_len_lat_lower_threshold=1e3,
+    grid_len_upper_threshold=1e5,
+    grid_len_lower_threshold=1e3,
     temporal_start=1,
     temporal_end=366,                            
     temporal_step=20,  
@@ -94,10 +92,8 @@ model = AdaSTEMRegressor(
     save_gridding_plot = True,
     ensemble_fold=10,
     min_ensemble_required=7,
-    grid_len_lon_upper_threshold=1e5,
-    grid_len_lon_lower_threshold=1e3,
-    grid_len_lat_upper_threshold=1e5,
-    grid_len_lat_lower_threshold=1e3,
+    grid_len_upper_threshold=1e5,
+    grid_len_lower_threshold=1e3,
     temporal_start=1,
     temporal_end=366,                            
     temporal_step=20,  
