### 2024-01-18 chenyangkang/stemflow#30 @chenyangkang
[Feature] Speed boost? Using Geo indexing dependency
As suggested during the JOSS review, I should probably use geo-indexing for, like, prediction problem.

This issue is to see of geopandas will speed up the indexing-related tasks.
