### 2024-01-18 e889f96
Attemp to add spherical indexing system
