### 2024-01-18 chenyangkang/stemflow#2 @chenyangkang
Still exploring the possibility of using geopandas as dependency #30
