### 2024-01-22 chenyangkang/stemflow#30 @chenyangkang
`stemflow` contains jitter and rotation operation to induce random griddings.

![image](https://github.com/chenyangkang/stemflow/assets/59206195/c745c62f-3ed7-48d7-8094-3685537b1a2f)
