+++ stemflow/gridding/QuadGrid.py
@@ -13,7 +13,8 @@ import pandas as pd
 
 from ..utils.generate_soft_colors import generate_soft_color
 from ..utils.validation import check_random_state
-from .Q_blocks import Grid, Node, Point
+from .Q_blocks import QGrid, QNode, QPoint
+from ..utils.jitterrotation.jitterrotator import JitterRotator
 
 os.environ["MKL_NUM_THREADS"] = "1"
 os.environ["NUMEXPR_NUM_THREADS"] = "1"
@@ -64,7 +65,7 @@ class QuadGrid:
         lat_new = (data[:, 1] + self.calibration_point_y_jitter).tolist()
 
         for index, lon, lat in zip(indexes, lon_new, lat_new):
-            self.points.append(Point(index, lon, lat))
+            self.points.append(QPoint(index, lon, lat))
 
     def generate_gridding_params(self):
         """For completeness"""
@@ -85,10 +86,10 @@ class QuadGrid:
         ymin = np.min(y_list)
         ymax = np.max(y_list)
 
-        self.x_start = xmin - self.grid_len + self.calibration_point_x_jitter
-        self.x_end = xmax + self.grid_len + self.calibration_point_x_jitter
-        self.y_start = ymin - self.grid_len + self.calibration_point_y_jitter
-        self.y_end = ymax + self.grid_len + self.calibration_point_y_jitter
+        self.x_start = xmin - self.grid_len
+        self.x_end = xmax + self.grid_len
+        self.y_start = ymin - self.grid_len
+        self.y_end = ymax + self.grid_len
         x_grids = np.arange(self.x_start, self.x_end, self.grid_len)
         y_grids = np.arange(self.y_start, self.y_end, self.grid_len)
 
@@ -99,7 +100,7 @@ class QuadGrid:
         self.grids = []
         for i in range(len(x_grids) - 1):
             for j in range(len(y_grids) - 1):
-                gird = Grid(i, j, (x_grids[i], x_grids[i + 1]), (y_grids[j], y_grids[j + 1]))
+                gird = QGrid(i, j, (x_grids[i], x_grids[i + 1]), (y_grids[j], y_grids[j + 1]))
                 self.grids.append(gird)
 
         # Use numpy.digitize to bin points into grids
@@ -112,30 +113,25 @@ class QuadGrid:
             grid.points = [self.points[i] for i in indices]
 
     def graph(self, scatter: bool = True, ax=None):
+        """plot gridding
+
+        Args:
+            scatter: Whether add scatterplot of data points
+        """
+        
         the_color = generate_soft_color()
 
-        theta = -(self.rotation_angle / 360) * np.pi * 2
-        rotation_matrix = np.array([[np.cos(theta), -np.sin(theta)], [np.sin(theta), np.cos(theta)]])
-
         for grid in self.grids:
-            xy0_trans = np.array([[grid.x_range[0], grid.y_range[0]]])
-            if self.calibration_point_x_jitter:
-                new_x = xy0_trans[:, 0] - self.calibration_point_x_jitter
-            else:
-                new_x = xy0_trans[:, 0]
 
-            if self.calibration_point_y_jitter:
-                new_y = xy0_trans[:, 1] - self.calibration_point_y_jitter
-            else:
-                new_y = xy0_trans[:, 1]
-            new_xy = np.array([[new_x[0], new_y[0]]]) @ rotation_matrix
-            new_x = new_xy[:, 0]
-            new_y = new_xy[:, 1]
+            old_x, old_y = JitterRotator.inverse_jitter_rotate([grid.x_range[0]], [grid.y_range[0]], 
+                                                self.rotation_angle,
+                                                self.calibration_point_x_jitter,
+                                                self.calibration_point_y_jitter)
 
             if ax is None:
[…]
