+++ stemflow/model/static_func_AdaSTEM.py
@@ -18,25 +18,9 @@ from sklearn.base import BaseEstimator
 from sklearn.utils import class_weight
 
 from .dummy_model import dummy_model1
+from ..utils.jitterrotation.jitterrotator import JitterRotator
 
 # warnings.filterwarnings("ignore")
-
-
-def _monkey_patched_predict_proba(
-    model: BaseEstimator, X_train: Union[pd.core.frame.DataFrame, np.ndarray]
-) -> np.ndarray:
-    """the monkey patching predict_proba method
-
-    Args:
-        model: the input model
-        X_train: input training data
-
-    Returns:
-        predicted proba
-    """
-    pred = model.predict(X_train)
-    pred = np.array(pred).reshape(-1, 1)
-    return np.concatenate([np.zeros(shape=pred.shape), pred], axis=1)
 
 
 def train_one_stixel(
@@ -46,8 +30,7 @@ def train_one_stixel(
     base_model: BaseEstimator,
     sample_weights_for_classifier: bool,
     subset_x_names: bool,
-    X_train_copy: pd.core.frame.DataFrame,
-    checklist_indexes: list,
+    stixel_X_train: pd.core.frame.DataFrame,
 ) -> Tuple[Union[None, BaseEstimator], list]:
     """Train one stixel
 
@@ -58,30 +41,28 @@ def train_one_stixel(
         base_model (BaseEstimator): Base model estimator.
         sample_weights_for_classifier (bool): Whether to balance the sample weights in classifier for imbalanced samples.
         subset_x_names (bool): Whether to only store variables with std > 0 for each stixel.
-        X_train_copy (pd.core.frame.DataFrame): Input training dataframe.
-        checklist_indexes (list): Each element is a list that contain all checklist indexes for this stixel.
+        sub_X_train (pd.core.frame.DataFrame): Input training dataframe for THE stixel.
 
     Returns:
         tuple[Union[None, BaseEstimator], list]: trained_model, stixel_specific_x_names
     """
-    sub_X_train = X_train_copy.iloc[checklist_indexes, :]
 
-    if len(sub_X_train) < stixel_training_size_threshold:  # threshold
-        return (None, [])
+    if len(stixel_X_train) < stixel_training_size_threshold:  # threshold
+        return (None, [], 'Not_Enough_Data')
 
-    sub_y_train = sub_X_train.iloc[:, -1]
-    sub_X_train = sub_X_train[x_names]
+    sub_y_train = stixel_X_train['true_y']
+    sub_X_train = stixel_X_train[x_names]
     unique_sub_y_train_binary = np.unique(np.where(sub_y_train > 0, 1, 0))
 
     # nan check
-    nan_count = np.sum(np.isnan(sub_y_train)) + np.sum(np.isnan(sub_y_train))
+    nan_count = np.sum(np.isnan(np.array(sub_X_train) ))+ np.sum(np.isnan(sub_y_train))
     if nan_count > 0:
-        return (None, [])
+        return (None, [], 'Contain_Nan')
 
     # fit
     if (not task == "regression") and (len(unique_sub_y_train_binary) == 1):
         trained_model = dummy_model1(float(unique_sub_y_train_binary[0]))
-        return (trained_model, [])
+        return (trained_model, [], 'Success')
     else:
         # Remove the variables that have no variation
         stixel_specific_x_names = x_names.copy()
@@ -93,7 +74,7 @@ def train_one_stixel(
 
         # continue, if no variable left
         if len(stixel_specific_x_names) == 0:
-            return (None, [])
+            return (None, [], 'x_names_length_zero')
 
[…]
