+++ stemflow/gridding/QTree.py
@@ -13,7 +13,8 @@ import pandas as pd
 
 from ..utils.generate_soft_colors import generate_soft_color
 from ..utils.validation import check_random_state
-from .Q_blocks import Node, Point
+from .Q_blocks import QNode, QPoint
+from ..utils.jitterrotation.jitterrotator import JitterRotator
 
 os.environ["MKL_NUM_THREADS"] = "1"
 os.environ["NUMEXPR_NUM_THREADS"] = "1"
@@ -23,7 +24,7 @@ warnings.filterwarnings("ignore")
 
 
 def recursive_subdivide(
-    node: Node,
+    node: QNode,
     grid_len_lon_upper_threshold: Union[float, int],
     grid_len_lon_lower_threshold: Union[float, int],
     grid_len_lat_upper_threshold: Union[float, int],
@@ -53,7 +54,7 @@ def recursive_subdivide(
     h_ = float(node.height / 2)
 
     p = contains(node.x0, node.y0, w_, h_, node.points)
-    x1 = Node(node.x0, node.y0, w_, h_, p)
+    x1 = QNode(node.x0, node.y0, w_, h_, p)
     recursive_subdivide(
         x1,
         grid_len_lon_upper_threshold,
@@ -64,7 +65,7 @@ def recursive_subdivide(
     )
 
     p = contains(node.x0, node.y0 + h_, w_, h_, node.points)
-    x2 = Node(node.x0, node.y0 + h_, w_, h_, p)
+    x2 = QNode(node.x0, node.y0 + h_, w_, h_, p)
     recursive_subdivide(
         x2,
         grid_len_lon_upper_threshold,
@@ -75,7 +76,7 @@ def recursive_subdivide(
     )
 
     p = contains(node.x0 + w_, node.y0, w_, h_, node.points)
-    x3 = Node(node.x0 + w_, node.y0, w_, h_, p)
+    x3 = QNode(node.x0 + w_, node.y0, w_, h_, p)
     recursive_subdivide(
         x3,
         grid_len_lon_upper_threshold,
@@ -86,7 +87,7 @@ def recursive_subdivide(
     )
 
     p = contains(node.x0 + w_, node.y0 + h_, w_, h_, node.points)
-    x4 = Node(node.x0 + w_, node.y0 + h_, w_, h_, p)
+    x4 = QNode(node.x0 + w_, node.y0 + h_, w_, h_, p)
     recursive_subdivide(
         x4,
         grid_len_lon_upper_threshold,
@@ -203,18 +204,14 @@ class QTree:
         """
         if not len(x_array) == len(y_array) or not len(x_array) == len(indexes):
             raise ValueError("input longitude and latitude and indexes not in same length!")
-
-        data = np.array([x_array, y_array]).T
-        angle = self.rotation_angle
-        r = angle / 360
-        theta = r * np.pi * 2
-        rotation_matrix = np.array([[np.cos(theta), -np.sin(theta)], [np.sin(theta), np.cos(theta)]])
-        data = data @ rotation_matrix
-        lon_new = (data[:, 0] + self.calibration_point_x_jitter).tolist()
-        lat_new = (data[:, 1] + self.calibration_point_y_jitter).tolist()
-
+        
+        lon_new, lat_new = JitterRotator.rotate_jitter(x_array, y_array, 
+                                                       self.rotation_angle, 
+                                                       self.calibration_point_x_jitter,
+                                                       self.calibration_point_y_jitter)
+        
         for index, lon, lat in zip(indexes, lon_new, lat_new):
-            self.points.append(Point(index, lon, lat))
+            self.points.append(QPoint(index, lon, lat))
 
     def generate_gridding_params(self):
         """generate the gridding params after data are added
[…]

[…]
