+++ stemflow/model/__init__.py
@@ -1,5 +1,1 @@
-from .static_func_AdaSTEM import (
-    assign_points_to_one_ensemble,
-    train_one_stixel,
-    transform_pred_set_to_STEM_quad,
-)
+from .static_func_AdaSTEM import assign_points_to_one_ensemble, train_one_stixel, transform_pred_set_to_STEM_quad
