+++ stemflow/utils/validation.py
@@ -1,7 +1,9 @@
-from typing import Union
 import warnings
+from typing import Union
+
 import numpy as np
 import pandas as pd
+
 
 def check_random_state(seed: Union[None, int, np.random.RandomState]) -> np.random.RandomState:
     """Turn seed into a np.random.RandomState instance.
@@ -24,16 +26,16 @@ def check_random_state(seed: Union[None, int, np.random.RandomState]) -> np.rand
         return seed
     raise ValueError("%r cannot be used to seed a numpy.random.RandomState instance" % seed)
 
+
 def check_task(task):
     if task not in ["regression", "classification", "hurdle"]:
-        raise AttributeError(
-            f"task type must be one of 'regression', 'classification', or 'hurdle'! Now it is {task}"
-        )
+        raise AttributeError(f"task type must be one of 'regression', 'classification', or 'hurdle'! Now it is {task}")
     if task == "hurdle":
         warnings.warn(
             "You have chosen HURDLE task. The goal is to first conduct classification, and then apply regression on points with *positive values*"
         )
-        
+
+
 def check_base_model(base_model):
     for func in ["fit", "predict"]:
         if func not in dir(base_model):
@@ -47,12 +49,13 @@ def check_njobs(njobs):
     elif njobs > 1:
         raise NotImplementedError("Multi-thread processing is not implemented yet.")
 
-        my_cpu_count = cpu_count()
-        if njobs > my_cpu_count:
-            raise ValueError(f"Setting of njobs ({njobs}) exceed the maximum ({my_cpu_count}).")
+        # my_cpu_count = cpu_count()
+        # if njobs > my_cpu_count:
+        #     raise ValueError(f"Setting of njobs ({njobs}) exceed the maximum ({my_cpu_count}).")
     else:
         pass
-    
+
+
 def check_verbosity(self, verbosity):
     if verbosity is None:
         verbosity = self.verbosity
@@ -61,32 +64,29 @@ def check_verbosity(self, verbosity):
     else:
         verbosity = 1
     return verbosity
-       
- 
+
+
 def check_spatio_bin_jitter_magnitude(spatio_bin_jitter_magnitude):
     if isinstance(spatio_bin_jitter_magnitude, (int, float)):
         pass
     elif isinstance(spatio_bin_jitter_magnitude, str):
-        if spatio_bin_jitter_magnitude == 'adaptive':
+        if spatio_bin_jitter_magnitude == "adaptive":
             pass
         else:
-            raise ValueError('spatio_bin_jitter_magnitude string must be adaptive!')
+            raise ValueError("spatio_bin_jitter_magnitude string must be adaptive!")
     else:
-        raise ValueError('spatio_bin_jitter_magnitude string must be one of [int, float, \'adaptive\']!')
-            
-            
+        raise ValueError("spatio_bin_jitter_magnitude string must be one of [int, float, 'adaptive']!")
+
+
 def check_transform_spatio_bin_jitter_magnitude(data, Spatio1, Spatio2, spatio_bin_jitter_magnitude):
     check_spatio_bin_jitter_magnitude(spatio_bin_jitter_magnitude)
     if isinstance(spatio_bin_jitter_magnitude, str):
-        if spatio_bin_jitter_magnitude=='adaptive':
-            jit = max(
-                data[Spatio1].max() - data[Spatio1].min(),
-                data[Spatio2].max() - data[Spatio2].min()
-            )
+        if spatio_bin_jitter_magnitude == "adaptive":
+            jit = max(data[Spatio1].max() - data[Spatio1].min(), data[Spatio2].max() - data[Spatio2].min())
             return jit
     return spatio_bin_jitter_magnitude
-                       
[…]
