+++ stemflow/model/STEM.py
@@ -33,7 +33,7 @@ class STEM(AdaSTEM):
         temporal_step: Union[float, int] = 20,
         temporal_bin_interval: Union[float, int] = 50,
         temporal_bin_start_jitter: Union[float, int, str] = "adaptive",
-        spatio_bin_jitter_magnitude: Union[float, int] = 'adaptive',
+        spatio_bin_jitter_magnitude: Union[float, int] = "adaptive",
         save_gridding_plot: bool = True,
         save_tmp: bool = False,
         save_dir: str = "./",
@@ -216,7 +216,7 @@ class STEMClassifier(AdaSTEMClassifier):
         temporal_step=20,
         temporal_bin_interval=50,
         temporal_bin_start_jitter="adaptive",
-        spatio_bin_jitter_magnitude='adaptive',
+        spatio_bin_jitter_magnitude="adaptive",
         save_gridding_plot=False,
         save_tmp=False,
         save_dir="./",
@@ -305,7 +305,7 @@ class STEMRegressor(AdaSTEMRegressor):
         temporal_step=20,
         temporal_bin_interval=50,
         temporal_bin_start_jitter="adaptive",
-        spatio_bin_jitter_magnitude='adaptive',
+        spatio_bin_jitter_magnitude="adaptive",
         save_gridding_plot=False,
         save_tmp=False,
         save_dir="./",
