+++ stemflow/model/AdaSTEM.py
@@ -32,26 +32,23 @@ from tqdm.auto import tqdm as tqdm_auto
 
 #
 from ..utils.quadtree import get_ensemble_quadtree
-from .dummy_model import dummy_model1
 from ..utils.validation import (
-    check_random_state, 
-    check_X_train, 
-    check_y_train, 
-    check_verbosity,
-    check_spatio_bin_jitter_magnitude,
-    check_temporal_bin_start_jitter,
-    check_task,
-    check_njobs,
     check_base_model,
-    check_X_test,
+    check_njobs,
     check_prediciton_aggregation,
-    check_prediction_return
+    check_prediction_return,
+    check_random_state,
+    check_spatio_bin_jitter_magnitude,
+    check_task,
+    check_temporal_bin_start_jitter,
+    check_verbosity,
+    check_X_test,
+    check_X_train,
+    check_y_train,
 )
-
-from ..utils.wrapper import (
-    model_wrapper
-)
-
+from ..utils.wrapper import model_wrapper
+from .dummy_model import dummy_model1
+from .Hurdle import Hurdle
 from .static_func_AdaSTEM import (  # predict_one_ensemble
     assign_points_to_one_ensemble,
     get_model_and_stixel_specific_x_names,
@@ -60,7 +57,6 @@ from .static_func_AdaSTEM import (  # predict_one_ensemble
     transform_pred_set_to_STEM_quad,
 )
 
-from .Hurdle import Hurdle
 #
 
 
@@ -85,7 +81,7 @@ class AdaSTEM(BaseEstimator):
         temporal_step: Union[float, int] = 20,
         temporal_bin_interval: Union[float, int] = 50,
         temporal_bin_start_jitter: Union[float, int, str] = "adaptive",
-        spatio_bin_jitter_magnitude: Union[float, int] = 'adaptive',
+        spatio_bin_jitter_magnitude: Union[float, int] = "adaptive",
         save_gridding_plot: bool = True,
         save_tmp: bool = False,
         save_dir: str = "./",
@@ -201,14 +197,14 @@ class AdaSTEM(BaseEstimator):
         check_base_model(base_model)
         base_model = model_wrapper(base_model)
         self.base_model = base_model
-        
+
         # 2. Model params
         check_task(task)
         self.task = task
         self.Temporal1 = Temporal1
         self.Spatio1 = Spatio1
         self.Spatio2 = Spatio2
-        
+
         # 3. Gridding params
         self.ensemble_fold = ensemble_fold
         self.min_ensemble_required = min_ensemble_required
@@ -228,7 +224,7 @@ class AdaSTEM(BaseEstimator):
         self.spatio_bin_jitter_magnitude = spatio_bin_jitter_magnitude
         check_temporal_bin_start_jitter(temporal_bin_start_jitter)
         self.temporal_bin_start_jitter = temporal_bin_start_jitter
-        
+
         # 4. Training params
         if stixel_training_size_threshold is None:
             self.stixel_training_size_threshold = points_lower_threshold
@@ -237,24 +233,24 @@ class AdaSTEM(BaseEstimator):
         self.use_temporal_to_train = use_temporal_to_train
         self.subset_x_names = subset_x_names
         self.sample_weights_for_classifier = sample_weights_for_classifier
-        
+
         # 5. Multi-threading params (not implemented yet)
         check_njobs(njobs)
         self.njobs = njobs
-        
+
         # 6. Plotting params
         self.plot_xlims = plot_xlims
         self.plot_ylims = plot_ylims
         self.save_tmp = save_tmp
         self.save_dir = save_dir
         self.save_gridding_plot = save_gridding_plot
-        
+
         # X. miscellaneous
[…]
