+++ stemflow/manually_testing.py
@@ -380,7 +380,6 @@ def run_mini_test(
     assert os.path.exists(os.path.join(tmp_dir, "error_plot.pdf"))
 
     # 11.Evaluation
-
     # %%
     print("Predicting on test set...")
     pred = model.predict(X_test)
@@ -389,7 +388,7 @@ def run_mini_test(
     # %%
     perc = np.sum(np.isnan(pred.flatten())) / len(pred.flatten())
     print(f"Percentage not predictable {round(perc*100, 2)}%")
-    assert perc < 0.05
+    assert perc < 0.5
 
     # %%
     pred_df = pd.DataFrame(
