+++ stemflow/utils/jitterrotation/jitterrotator.py
@@ -121,3 +121,90 @@ class JitterRotator:
         ).T
         back_rotated = back_jitter_data @ rotation_matrix
         return back_rotated[:, 0].flatten(), back_rotated[:, 1].flatten()
+
+
+class Sphere_Jitterrotator:
+    def __init__(self) -> None:
+        pass
+
+    def rotate_jitter(point: np.ndarray, axis: np.ndarray, angle: Union[float, int]):
+        """_summary_
+
+        Args:
+            point (np.ndarray): shape of (X, 3)
+            axis (np.ndarray): shape of (3,)
+            angle (Union[float, int]): angle in degree
+
+        Returns:
+            _type_: _description_
+        """
+        u = np.array(axis)
+        u = u / np.linalg.norm(u)
+
+        angle_ = angle * (np.pi / 180)
+        cos_theta = np.cos(angle_)
+        sin_theta = np.sin(angle_)
+
+        rotation_matrix = np.array(
+            [
+                [
+                    cos_theta + u[0] ** 2 * (1 - cos_theta),
+                    u[0] * u[1] * (1 - cos_theta) - u[2] * sin_theta,
+                    u[0] * u[2] * (1 - cos_theta) + u[1] * sin_theta,
+                ],
+                [
+                    u[1] * u[0] * (1 - cos_theta) + u[2] * sin_theta,
+                    cos_theta + u[1] ** 2 * (1 - cos_theta),
+                    u[1] * u[2] * (1 - cos_theta) - u[0] * sin_theta,
+                ],
+                [
+                    u[2] * u[0] * (1 - cos_theta) - u[1] * sin_theta,
+                    u[2] * u[1] * (1 - cos_theta) + u[0] * sin_theta,
+                    cos_theta + u[2] ** 2 * (1 - cos_theta),
+                ],
+            ]
+        )
+
+        rotated_point = np.dot(point, rotation_matrix)
+        return rotated_point
+
+    def inverse_rotate_jitter(point: np.ndarray, axis: np.ndarray, angle: Union[float, int]):
+        """_summary_
+
+        Args:
+            point (np.ndarray): shape of (X, 3)
+            axis (np.ndarray): shape of (3,)
+            angle (Union[float, int]): angle in degree
+
+        Returns:
+            _type_: _description_
+        """
+        u = np.array(axis)
+        u = u / np.linalg.norm(u)
+
+        angle_ = -angle * (np.pi / 180)
+        cos_theta = np.cos(angle_)
+        sin_theta = np.sin(angle_)
+
+        rotation_matrix = np.array(
+            [
+                [
+                    cos_theta + u[0] ** 2 * (1 - cos_theta),
+                    u[0] * u[1] * (1 - cos_theta) - u[2] * sin_theta,
+                    u[0] * u[2] * (1 - cos_theta) + u[1] * sin_theta,
+                ],
+                [
+                    u[1] * u[0] * (1 - cos_theta) + u[2] * sin_theta,
+                    cos_theta + u[1] ** 2 * (1 - cos_theta),
+                    u[1] * u[2] * (1 - cos_theta) - u[0] * sin_theta,
+                ],
+                [
+                    u[2] * u[0] * (1 - cos_theta) - u[1] * sin_theta,
+                    u[2] * u[1] * (1 - cos_theta) + u[0] * sin_theta,
+                    cos_theta + u[2] ** 2 * (1 - cos_theta),
+                ],
+            ]
+        )
+
+        rotated_point = np.dot(point, rotation_matrix)
+        return rotated_point
