### 2024-01-23 40b3e5b
Add SphereAdaSTEM modules and make mini test
