+++ tests/make_models.py
@@ -144,7 +144,8 @@ def make_SphereAdaSTEMRegressor(
         save_gridding_plot=True,
         ensemble_fold=fold_,
         min_ensemble_required=min_req,
-        grid_len=30,
+        grid_len_upper_threshold=5000,
+        grid_len_lower_threshold=100,
         temporal_start=1,
         temporal_end=366,
         temporal_step=30,
@@ -169,7 +170,8 @@ def make_SphereAdaClassifier(fold_=2, min_req=1, ensemble_models_disk_saver=Fals
         save_gridding_plot=True,
         ensemble_fold=fold_,
         min_ensemble_required=min_req,
-        grid_len=30,
+        grid_len_upper_threshold=5000,
+        grid_len_lower_threshold=100,
         temporal_start=1,
         temporal_end=366,
         temporal_step=30,
