+++ stemflow/utils/plot_gif.py
@@ -101,8 +101,8 @@ def make_sample_gif(
         sub = data[data["Temporal_indexer"] == i]
         temporal_value = np.array(sub[Temporal1].values)[0]
 
-        sub[f"{Spatio1}_grid"] = np.digitize(sub[Spatio1], lng_gird, right=True)
-        sub[f"{Spatio2}_grid"] = np.digitize(sub[Spatio2], lat_gird, right=False)
+        sub.loc[:, f"{Spatio1}_grid"] = np.digitize(sub[Spatio1], lng_gird, right=True)
+        sub.loc[:, f"{Spatio2}_grid"] = np.digitize(sub[Spatio2], lat_gird, right=False)
         sub = sub[(sub[f"{Spatio1}_grid"] <= lng_size - 1) & (sub[f"{Spatio2}_grid"] <= lat_size - 1)]
 
         sub = sub.groupby([f"{Spatio1}_grid", f"{Spatio2}_grid"])[[col]].mean().reset_index(drop=False)
@@ -114,7 +114,7 @@ def make_sample_gif(
         else:
             im[sub[f"{Spatio2}_grid"].values, sub[f"{Spatio1}_grid"].values] = sub[col]
 
-        my_cmap = matplotlib.cm.get_cmap(cmap)
+        my_cmap = matplotlib.colormaps.get_cmap(cmap)
 
         if lightgrey_under:
             my_cmap.set_under("lightgrey")
