+++ stemflow/model/static_func_AdaSTEM.py
@@ -222,9 +222,9 @@ def assign_points_to_one_ensemble_sphere(
 
     def find_belonged_points(df, df_a):
         P0 = np.array([0, 0, 0]).reshape(1, -1)
-        A = np.array(df[["p1x", "p1y", "p1z"]].iloc[0])
-        B = np.array(df[["p2x", "p2y", "p2z"]].iloc[0])
-        C = np.array(df[["p3x", "p3y", "p3z"]].iloc[0])
+        A = np.array(df[["p1x", "p1y", "p1z"]].values)
+        B = np.array(df[["p2x", "p2y", "p2z"]].values)
+        C = np.array(df[["p3x", "p3y", "p3z"]].values)
 
         intersect = intersect_triangle_plane(
             P0=P0, V=df_a[["x_3D_transformed", "y_3D_transformed", "z_3D_transformed"]].values, A=A, B=B, C=C
@@ -440,7 +440,7 @@ def predict_one_stixel(
     """predict_one_stixel
 
     Args:
-        X_test_copy (pd.core.frame.DataFrame): Input testing variables
+        X_test_stixel (pd.core.frame.DataFrame): Input testing variables
         task (str): One of 'regression', 'classification' and 'hurdle'
         model_x_names_tuple (tuple[Union[None, BaseEstimator], list]): A tuple of (model, stixel_specific_x_names)
 
