+++ requirements.txt
@@ -3,6 +3,7 @@
 matplotlib>=3.7.1
 numpy>=1.24.3
 pandas>=1.5.3
+plotly>=5.9.0
 scikit_learn>=1.2.2
 # watermark>=2.3.1
 # xgboost>=1.7.6
