### 2024-01-24 0974836
fix test bug and add docstrings
