+++ stemflow/utils/sphere/discriminant_formula.py
@@ -1,21 +1,20 @@
+from typing import Union
+
 import numpy as np
 
-# def sign(target, p2, p3):
-#     return np.sign((target[:,0] - p3[0]) * (p2[1] - p3[1]) - (p2[0] - p3[0]) * (target[:,1] - p3[1]))
 
-# def point_in_triangle(targets, p1, p2, p3):
+def is_point_inside_triangle(point: np.ndarray, A: np.ndarray, B: np.ndarray, C: np.ndarray) -> np.ndarray:
+    """Check if a point is inside a triangle
 
-#     d1 = sign(targets, p1, p2)
-#     d2 = sign(targets, p2, p3)
-#     d3 = sign(targets, p3, p1)
+    Args:
+        point (np.ndarray): point in vector. Shape (X, dimension).
+        A (np.ndarray): point A of triangle. Shape (dimension).
+        B (np.ndarray): point B of triangle. Shape (dimension).
+        C (np.ndarray): point C of triangle. Shape (dimension).
 
-#     signs = np.column_stack([d1<0,d2<0,d3<0])
-#     has_neg = signs.sum(axis=1)
-#     has_pos = -signs.sum(axis=1)
-#     return np.logical_not(np.logical_and(has_neg, has_pos))
-
-
-def is_point_inside_triangle(point, A, B, C):
+    Returns:
+        np.ndarray: inside or not
+    """
     u = np.cross(C - B, point - B) @ np.cross(C - B, A - B)
     v = np.cross(A - C, point - C) @ np.cross(A - C, B - C)
     w = np.cross(B - A, point - A) @ np.cross(B - A, C - A)
@@ -23,7 +22,19 @@ def is_point_inside_triangle(point, A, B, C):
     return (u >= 0) & (v >= 0) & (w >= 0)
 
 
-def intersect_triangle_plane(P0, V, A, B, C):
+def intersect_triangle_plane(P0: np.ndarray, V: np.ndarray, A: np.ndarray, B: np.ndarray, C: np.ndarray) -> np.ndarray:
+    """Get if the ray go through the triangle of A,B,C
+
+    Args:
+        P0 (np.ndarray): start point of ray
+        V (np.ndarray): A point that the ray go through
+        A (np.ndarray): point A of triangle. Shape (dimension).
+        B (np.ndarray): point A of triangle. Shape (dimension).
+        C (np.ndarray): point A of triangle. Shape (dimension).
+
+    Returns:
+        np.ndarray: Whether the point go through triangle ABC
+    """
     # Calculate the normal vector of the plane
     N = np.cross(B - A, C - A)
 
