### 2024-01-24 1607e44
fix test bug and add docstrings
