+++ stemflow/gridding/Sphere_QTree.py
@@ -12,6 +12,7 @@ import matplotlib.pyplot as plt  # plotting libraries
 import numpy as np
 import pandas
 import pandas as pd
+import plotly.express as px
 from tqdm import tqdm
 
 from ..utils.generate_soft_colors import generate_soft_color
@@ -36,7 +37,7 @@ os.environ["OMP_NUM_THREADS"] = "1"
 warnings.filterwarnings("ignore")
 
 
-def recursive_subdivide(
+def Sphere_recursive_subdivide(
     node: Sphere_QTriangle,
     grid_len_upper_threshold: Union[float, int],
     grid_len_lower_threshold: Union[float, int],
@@ -55,7 +56,7 @@ def recursive_subdivide(
         points_lower_threshold:
             Stop splitting if fall short
         radius:
-            radius
+            radius of earth.
 
     """
 
@@ -74,7 +75,7 @@ def recursive_subdivide(
     pm23 = get_midpoint_3D(node.p2, node.p3, radius)
 
     # 1.
-    points_contained = contains(node.points, pm12, pm23, node.p2)
+    points_contained = Sphere_contains(node.points, pm12, pm23, node.p2)
     x1 = Sphere_QTriangle(
         pm12,
         pm23,
@@ -83,7 +84,7 @@ def recursive_subdivide(
         distance_from_3D_point(pm12.x, pm12.y, pm12.z, pm23.x, pm23.y, pm23.z, radius),
         radius,
     )
-    recursive_subdivide(
+    Sphere_recursive_subdivide(
         x1,
         grid_len_upper_threshold,
         grid_len_lower_threshold,
@@ -91,7 +92,7 @@ def recursive_subdivide(
     )
 
     # 2.
-    points_contained = contains(node.points, pm12, pm13, node.p1)
+    points_contained = Sphere_contains(node.points, pm12, pm13, node.p1)
     x2 = Sphere_QTriangle(
         pm12,
         pm13,
@@ -100,7 +101,7 @@ def recursive_subdivide(
         distance_from_3D_point(pm12.x, pm12.y, pm12.z, pm13.x, pm13.y, pm13.z, radius),
         radius,
     )
-    recursive_subdivide(
+    Sphere_recursive_subdivide(
         x2,
         grid_len_upper_threshold,
         grid_len_lower_threshold,
@@ -108,7 +109,7 @@ def recursive_subdivide(
     )
 
     # 3.
-    points_contained = contains(node.points, pm23, pm13, node.p3)
+    points_contained = Sphere_contains(node.points, pm23, pm13, node.p3)
     x3 = Sphere_QTriangle(
         pm23,
         pm13,
@@ -117,7 +118,7 @@ def recursive_subdivide(
         distance_from_3D_point(pm13.x, pm13.y, pm13.z, pm23.x, pm23.y, pm23.z, radius),
         radius,
     )
-    recursive_subdivide(
+    Sphere_recursive_subdivide(
         x3,
         grid_len_upper_threshold,
         grid_len_lower_threshold,
@@ -125,7 +126,7 @@ def recursive_subdivide(
     )
 
     # 3.
-    points_contained = contains(node.points, pm12, pm13, pm23)
+    points_contained = Sphere_contains(node.points, pm12, pm13, pm23)
     x4 = Sphere_QTriangle(
         pm12,
         pm13,
@@ -134,7 +135,7 @@ def recursive_subdivide(
         distance_from_3D_point(pm12.x, pm12.y, pm12.z, pm13.x, pm13.y, pm13.z, radius),
         radius,
     )
-    recursive_subdivide(
+    Sphere_recursive_subdivide(
         x4,
         grid_len_upper_threshold,
[…]
