+++ stemflow/model/SphereAdaSTEM.py
@@ -1,36 +1,17 @@
 import os
-import pickle
 import warnings
-from itertools import repeat
-
-#
-from multiprocessing import Pool, cpu_count
-from typing import Tuple, Union
+from functools import partial
+from typing import Callable, Tuple, Union
 
 import numpy as np
 import pandas as pd
 from numpy import ndarray
 
 # validation check
-from pandas.core.frame import DataFrame
-from scipy.stats import pearsonr, spearmanr
 from sklearn.base import BaseEstimator
-from sklearn.metrics import (
-    average_precision_score,
-    cohen_kappa_score,
-    d2_tweedie_score,
-    f1_score,
-    mean_absolute_error,
-    mean_squared_error,
-    precision_score,
-    r2_score,
-    recall_score,
-    roc_auc_score,
-)
 from tqdm import tqdm
 from tqdm.auto import tqdm as tqdm_auto
 
-from ..utils.sphere.coordinate_transform import lonlat_cartesian_3D_transformer
 from ..utils.sphere.discriminant_formula import intersect_triangle_plane
 
 #
@@ -39,25 +20,15 @@ from ..utils.validation import (
     check_base_model,
     check_njobs,
     check_prediciton_aggregation,
-    check_prediction_return,
-    check_random_state,
     check_spatio_bin_jitter_magnitude,
     check_task,
     check_temporal_bin_start_jitter,
     check_verbosity,
-    check_X_test,
-    check_X_train,
-    check_y_train,
 )
 from ..utils.wrapper import model_wrapper
-from .AdaSTEM import AdaSTEM, AdaSTEMClassifier, AdaSTEMRegressor
-from .dummy_model import dummy_model1
-from .Hurdle import Hurdle
+from .AdaSTEM import AdaSTEM
 from .static_func_AdaSTEM import (  # predict_one_ensemble
     assign_points_to_one_ensemble_sphere,
-    get_model_and_stixel_specific_x_names,
-    predict_one_stixel,
-    train_one_stixel,
     transform_pred_set_to_Sphere_STEM_quad,
 )
 
@@ -108,7 +79,7 @@ class SphereAdaSTEM(AdaSTEM):
         plot_ylims: Tuple[Union[float, int], Union[float, int]] = (-90, 90),
         verbosity: int = 0,
     ):
-        """Make an AdaSTEM object
+        """Make a Spherical AdaSTEM object
 
         Args:
             base_model:
@@ -122,9 +93,9 @@ class SphereAdaSTEM(AdaSTEM):
                 In the training phase, if stixels contain less than `points_lower_threshold` of data records,
                 the results are set to np.nan, making them `unpredictable`. Defaults to 7.
             grid_len_upper_threshold:
-                force divide if grid length larger than the threshold. Defaults to 25.
+                force divide if grid length (km) larger than the threshold. Defaults to 8000 km.
             grid_len_lower_threshold:
-                stop divide if grid length **will** be below than the threshold. Defaults to 5.
+                stop divide if grid length (km) **will** be below than the threshold. Defaults to 500 km.
             points_lower_threshold:
                 Do not further split the gird if split results in less samples than this threshold.
                 Overriden by grid_len_*_upper_threshold parameters. Defaults to 50.
@@ -315,7 +286,15 @@ class SphereAdaSTEM(AdaSTEM):
             ax=ax,
         )
 
-    def SAC_ensemble_training(self, index_df, data):
+    def SAC_ensemble_training(self, index_df: pd.core.frame.DataFrame, data: pd.core.frame.DataFrame):
+        """A sub-module of SAC training function.
+        Train only one ensemble.
+
+        Args:
+            index_df (pd.core.frame.DataFrame): ensemble data (model.ensemble_df)
[…]
