### 2024-01-24 a11c6c4
finish updating the docstrings #33
