+++ stemflow/model/AdaSTEM.py
@@ -5,7 +5,7 @@ from itertools import repeat
 
 #
 from multiprocessing import Pool, cpu_count
-from typing import Tuple, Union
+from typing import Callable, Tuple, Union
 
 import numpy as np
 import pandas as pd
@@ -304,7 +304,12 @@ class AdaSTEM(BaseEstimator):
             ax=ax,
         )
 
-    def store_x_names(self, X_train):
+    def store_x_names(self, X_train: pd.core.frame.DataFrame):
+        """Store the training variables
+
+        Args:
+            X_train (pd.core.frame.DataFrame): input training data.
+        """
         # store x_names
         self.x_names = list(X_train.columns)
         if not self.use_temporal_to_train:
@@ -319,10 +324,10 @@ class AdaSTEM(BaseEstimator):
                 del self.x_names[self.x_names.index(i)]
 
     def stixel_fitting(self, stixel):
-        """Fit one stixel
+        """A sub module of SAC training. Fit one stixel
 
         Args:
-            stixel (gpd.geodataframe.GeoDataFrame): data sjoined with ensemble_df.
+            stixel (pd.core.frame.DataFrame): data sjoined with ensemble_df.
             For a single stixel.
         """
 
@@ -348,7 +353,15 @@ class AdaSTEM(BaseEstimator):
             self.model_dict[f"{name}_model"] = model
             self.stixel_specific_x_names[name] = stixel_specific_x_names
 
-    def SAC_ensemble_training(self, index_df, data):
+    def SAC_ensemble_training(self, index_df: pd.core.frame.DataFrame, data: pd.core.frame.DataFrame):
+        """A sub-module of SAC training function.
+        Train only one ensemble.
+
+        Args:
+            index_df (pd.core.frame.DataFrame): ensemble data (model.ensemble_df)
+            data (pd.core.frame.DataFrame): input covariates to train
+        """
+
         # Calculate the start indices for the sliding window
         unique_start_indices = np.sort(index_df[f"{self.Temporal1}_start"].unique())
         # training, window by window
@@ -395,10 +408,15 @@ class AdaSTEM(BaseEstimator):
                 .apply(lambda stixel: self.stixel_fitting(stixel))
             )
 
-    def SAC_training(self, ensemble_df, data, verbosity):
-        """Training with the whole input data
+    def SAC_training(self, ensemble_df: pd.core.frame.DataFrame, data: pd.core.frame.DataFrame, verbosity: int = 0):
+        """This function is a training function with SAC strategy:
+        Split (S), Apply(A), Combine (C). At ensemble level.
+        It is built on pandas `apply` method.
 
-        split (S), apply(A), combine (C). Ensemble level
+        Args:
+            ensemble_df (pd.core.frame.DataFrame): gridding information for all ensemble
+            data (pd.core.frame.DataFrame): data
+            verbosity (int, optional): Defaults to 0.
 
         """
 
@@ -454,12 +472,15 @@ class AdaSTEM(BaseEstimator):
 
         return self
 
-    def stixel_predict(self, stixel):
-        """Predict one stixel
+    def stixel_predict(self, stixel: pd.core.frame.DataFrame) -> Union[None, pd.core.frame.DataFrame]:
+        """A sub module of SAC prediction. Predict one stixel
 
         Args:
             stixel (pd.core.frame.DataFrame): data joined with ensemble_df.
             For a single stixel.
+
+        Returns:
+            pd.core.frame.DataFrame: the prediction result of this stixel
         """
         ensemble_index = stixel["ensemble_index"].iloc[0]
         unique_stixel_id = stixel["unique_stixel_id"].iloc[0]
@@ -482,12 +503,17 @@ class AdaSTEM(BaseEstimator):
         else:
             return pred
 
[…]
