+++ docs/API_Documentation/gridding/stemflow.gridding.Sphere_QTree.md
@@ -1,0 +1,3 @@
+# stemflow.gridding.Sphere_QTree
+
+::: stemflow.gridding.Sphere_QTree
