+++ docs/API_Documentation/utils/sphere/stemflow.utils.sphere.Icosahedron.md
@@ -1,0 +1,5 @@
+# stemflow.utils.sphere.Icosahedron
+
+---
+::: stemflow.utils.sphere.Icosahedron
+---
