+++ stemflow/version.py
@@ -1,1 +1,1 @@
-__version__ = "1.0.9.4"
+__version__ = "1.0.9.5"
