+++ stemflow/model/SphereAdaSTEM.py
@@ -1,6 +1,7 @@
 import os
 import warnings
 from functools import partial
+from types import MethodType
 from typing import Callable, Tuple, Union
 
 import numpy as np
@@ -26,7 +27,7 @@ from ..utils.validation import (
     check_verbosity,
 )
 from ..utils.wrapper import model_wrapper
-from .AdaSTEM import AdaSTEM
+from .AdaSTEM import AdaSTEM, AdaSTEMClassifier, AdaSTEMRegressor
 from .static_func_AdaSTEM import (  # predict_one_ensemble
     assign_points_to_one_ensemble_sphere,
     transform_pred_set_to_Sphere_STEM_quad,
@@ -79,6 +80,7 @@ class SphereAdaSTEM(AdaSTEM):
         plot_xlims: Tuple[Union[float, int], Union[float, int]] = (-180, 180),
         plot_ylims: Tuple[Union[float, int], Union[float, int]] = (-90, 90),
         verbosity: int = 0,
+        plot_empty: bool = False,
     ):
         """Make a Spherical AdaSTEM object
 
@@ -148,7 +150,8 @@ class SphereAdaSTEM(AdaSTEM):
                 If save_gridding_plot=true, what is the ylims of the plot. Defaults to (-90,90).
             verbosity:
                 0 to output nothing and everything otherwise.
-
+            plot_empty:
+                Whether to plot the empty grid
 
         Raises:
             AttributeError: Base model do not have method 'fit' or 'predict'
@@ -175,70 +178,46 @@ class SphereAdaSTEM(AdaSTEM):
                 feature importance dataframe for each stixel.
 
         """
-        # 1. Base model
-        check_base_model(base_model)
-        base_model = model_wrapper(base_model)
-        self.base_model = base_model
+        # Init parent class
+        super().__init__(
+            base_model,
+            task,
+            ensemble_fold,
+            min_ensemble_required,
+            grid_len_upper_threshold,
+            grid_len_lower_threshold,
+            points_lower_threshold,
+            stixel_training_size_threshold,
+            temporal_start,
+            temporal_end,
+            temporal_step,
+            temporal_bin_interval,
+            temporal_bin_start_jitter,
+            spatio_bin_jitter_magnitude,
+            save_gridding_plot,
+            save_tmp,
+            save_dir,
+            sample_weights_for_classifier,
+            Spatio1,
+            Spatio2,
+            Temporal1,
+            use_temporal_to_train,
+            njobs,
+            subset_x_names,
+            ensemble_models_disk_saver,
+            ensemble_models_disk_saving_dir,
+            plot_xlims,
+            plot_ylims,
+            verbosity,
+            plot_empty,
+        )
 
-        # 2. Model params
-        check_task(task)
-        self.task = task
-        self.Temporal1 = Temporal1
-        self.Spatio1 = Spatio1
-        self.Spatio2 = Spatio2
-        if not ((self.Spatio1 == "longitude") and (self.Spatio2 == "latitude")):
-            raise ValueError("Spatio1 and Spatio2 must be longitude and latitude!")
-
-        # 3. Gridding params
-        self.ensemble_fold = ensemble_fold
-        self.min_ensemble_required = min_ensemble_required
-        self.grid_len_upper_threshold = (
-            self.grid_len_lon_upper_threshold
-        ) = self.grid_len_lat_upper_threshold = grid_len_upper_threshold
-        self.grid_len_lower_threshold = (
-            self.grid_len_lon_lower_threshold
-        ) = self.grid_len_lat_lower_threshold = grid_len_lower_threshold
-        self.points_lower_threshold = points_lower_threshold
-        self.temporal_start = temporal_start
-        self.temporal_end = temporal_end
-        self.temporal_step = temporal_step
[…]
