+++ mkdocs.yml
@@ -44,8 +44,8 @@ theme:
 
   font:
     text: Roboto
-  logo: logo.png
-  favicon: logo.png
+  logo: assets/logo.png
+  favicon: assets/logo.png
 
 plugins:
   - social
@@ -74,14 +74,15 @@ nav:
   - Home: index.md
   - 'A Brief Introduction': A_brief_introduction/A_brief_introduction.md
   - Examples:
-    - Examples/00.Mini_test.ipynb
     - Examples/01.AdaSTEM_demo.ipynb
+    - Examples/04.SphereAdaSTEM_demo.ipynb
     - Examples/02.AdaSTEM_learning_curve_analysis.ipynb
     - Examples/03.Binding_with_Maxent.ipynb
     - Examples/05.Hurdle_in_ada_or_ada_in_hurdle.ipynb
     - Examples/06.Base_model_choices.ipynb
     - Examples/07.Optimizing_stixel_size.ipynb
   - Tips:
+    - 'Tips for spatiotemporal indexing': Tips/Tips_for_spatiotemporal_indexing.md
     - 'Tips for data types': Tips/Tips_for_data_types.md
     - 'Tips for different tasks': Tips/Tips_for_different_tasks.md
   - API Documentation:

[…]
