### 2024-01-25 3c27a54
clean the remnant of manually_testing
