### 2024-01-25 76b410a
fix bugs in SphereAdaSTEM
