### 2024-01-25 950f10e
fix bugs in SphereAdaSTEM
