### 2024-01-25 8ab0a7e
adjust figure size
