### 2024-01-25 4b20485
adjust figure size
