+++ paper/md_paper/paper.bib
@@ -1,3 +1,17 @@
+
+@article{wickham2011split,
+  title={The split-apply-combine strategy for data analysis},
+  author={Wickham, Hadley},
+  journal={Journal of statistical software},
+  volume={40},
+  issn = {1}
+  pages={1--29},
+  year={2011},
+  doi = {10.18637/jss.v040.i01},
+  url = {https://www.jstatsoft.org/article/view/v040i01}
+}
+
+
 @article{fink_adaptive_2013,
 	title = {Adaptive {Spatio}-{Temporal} {Exploratory} {Models}: {Hemisphere}-wide species distributions from massively crowdsourced {eBird} data},
 	volume = {27},
