+++ paper/md_paper/stemflow.bib
@@ -17,7 +17,17 @@
 }
 
 
-
+@article{wickham2011split,
+  title={The split-apply-combine strategy for data analysis},
+  author={Wickham, Hadley},
+  journal={Journal of statistical software},
+  volume={40},
+  issn = {1}
+  pages={1--29},
+  year={2011},
+  doi = {10.18637/jss.v040.i01},
+  url = {https://www.jstatsoft.org/article/view/v040i01}
+}
 
 
 @article{anderson_elapid_2023,
