+++ paper/md_paper/paper.bib
@@ -4,7 +4,7 @@
   author={Wickham, Hadley},
   journal={Journal of statistical software},
   volume={40},
-  issn = {1}
+  issn = {1},
   pages={1--29},
   year={2011},
   doi = {10.18637/jss.v040.i01},
