### 2024-01-26 #6158 @chenyangkang
Hi @jedalong, thank you for the valuable suggestions! I agree with most of your suggestions, and here are my responses:

1. According to your suggestion, I implemented `SphereAdaSTEM` module for spherical indexing and gridding, to solve the distortion problem. The gridding parameters can now be input as real distance in km. For a demo visualization, see [here](https://chenyangkang.github.io/stemflow/assets/Sphere_gridding.html). <br>
Related issue: https://github.com/chenyangkang/stemflow/issues/24

2. Besides the 3D indexing, for 2D indexing:
- I implemented `STEM` module where user can input fixed grid size (1 parameter).
- For `AdaSTEM`, I reduces the parameters from 4 to 2 (2 is necessary because gird size of `AdaSTEM` is adapted to sample size by design). <br>
Related issue: https://github.com/chenyangkang/stemflow/issues/23

3. You suggested that user should be able to input spatial object (I suppose you are referring to [shapely](https://github.com/shapely/shapely) and related), which I philosophically agree with you. However, after conducting [some test here](https://github.com/chenyangkang/stemflow/issues/30), I found that spatial object operation is much slower than the original `numpy` operation, which leverages vectorization. This could be an issue as `stemflow` is positioned for big data modeling, and also for other reasons (I discussed them in the issue thread). Some other methods were considered, like [scikit-spatial](https://github.com/ajhynes7/scikit-spatial), but it is simply a wrapper of `numpy` operation and I think that is not helpful in our circumstances. Consequently, I think I will keep the methodology unchanged. Please do let me know if you believe there are better solutions or tools that I missed.<br>
Related issue: https://github.com/chenyangkang/stemflow/issues/30

4. All changes are reflected in home page, documentations, and notebooks. Documentations are modified based on your suggestions (for example, a sliding window illustration is added [here](https://chenyangkang.github.io/stemflow/Examples/01.AdaSTEM_demo.html)). Paper is revised according to your suggestions.<br>
Related issues:
- https://github.com/chenyangkang/stemflow/issues/25
- https://github.com/chenyangkang/stemflow/issues/22

Again, thank you for your suggestions and contribution to `stemflow` and the community! Please feel free to open issues or reply if you have further concerns.
