### 2024-01-27 d9300dd
attemp to add mp outside prediction loop
