### 2024-01-27 61f6cb3
attemp to add mp outside prediction loop
