+++ stemflow/model/AdaSTEM.py
@@ -10,7 +10,7 @@ from multiprocessing import Pool, cpu_count
 from typing import Callable, Tuple, Union
 
 #
-import dask.dataframe as dd
+# import dask.dataframe as dd
 import numpy as np
 import pandas as pd
 from numpy import ndarray
