+++ paper/md_paper/paper_old.md
@@ -1,68 +1,0 @@
----
-title: 'stemflow: A Python Package for Adaptive Spatio-Temporal Exploratory Model'
-tags:
- - Python
- - ecology
- - spatio-temporal big data
- - citizen science
- - modeling
- - seasonal migration
-authors:
- - given-names: Yangkang
-   surname: Chen
-   orchid: 0000-0003-4996-0881
-   affiliation: "1, 2"
- - given-names: Zhongru
-   surname: Gu
-   orchid: 0000-0002-0178-1107
-   affiliation: "1, 3"
- - given-names: Xiangjiang
-   surname: Zhan
-   corresponding: true
-   orchid: 0000-0002-4517-1626
-   affiliation: "1, 2, 3, 4"
-
-affiliations:
- - name: Key Laboratory of Animal Ecology and Conservation Biology, Institute of Zoology, Chinese Academy of Sciences, Beijing, China
-   index: 1
-
- - name: University of Chinese Academy of Sciences, Beijing, China
-   index: 2
-
- - name: Cardiff University-Institute of Zoology Joint Laboratory for Biocomplexity Research, Chinese Academy of Sciences, Beijing, China
-   index: 3
-
- - name: Center for Excellence in Animal Evolution and Genetics, Chinese Academy of Sciences, Kunming, China
-   index: 4
-
-date: 20 September 2023
-bibliography: "paper.bib"
-
----
-
-# Summary
-
-Stemflow is a user-friendly Python package for Adaptive Spatio-Temporal Exploratory Model (AdaSTEM [@fink_adaptive_2013]). AdaSTEM is a modeling framework that adopts "split-apply-combine" methodology [@wickham2011split] – it adaptively splits data into spatiotemporal grids, train models for each grid, and combines the models for ensemble prediction. Models in stemflow follow the style of scikit-learn BaseEstimator class [@pedregosa2011scikit]. It provides one-line model creation, fitting, prediction, and evaluation. It implements spatio-temporal train-test-split and cross-validation functions. After model training, feature importance could be evaluated with spatio-temporal dynamics. Stemflow also provides functions for visualizing ensembles structured in model training and generating GIF file for predicted results to animate the spatio-temporal movement of animal population.
-
-# Statement of need
-
-Spatio-temporal big data is an increasingly valuable resource for modern ecological studies [@farley_situating_2018]. A large amount of spatio-temporal big data is now derived from broad-scale surveys, such as citizen science projects [@dickinson_citizen_2010]. The intensity of survey activities grows rapidly as more people are involved in citizen science in recent years, resulted in exponential accumulation of observational data [@di_cecco_observing_2021; @sullivan_ebird_2014]. However, daily species observation records uploaded by non-professionals in citizen science program are known to have larger bias than professionally structured research, both in terms of data veracity and spatio-temporal balance of the datasets, which necessitates elaborate modeling methods to provide insights [@dickinson_citizen_2010; @farley_situating_2018].
-
[…]
