### 2024-01-29 c5a38ee
add checks and update docs
