### 2024-01-31 dad2185
branch 1: split: process_map, train: multiprocessing with shared memroy, predict: single core
