### 2024-01-31 5155e71
branch 1: split: process_map, train: multiprocessing with shared memroy, predict: single core
