+++ stemflow/utils/quadtree.py
@@ -1,7 +1,9 @@
 "A function module to get quadtree results for 2D indexing system. Returns ensemble_df and plotting axes."
 
+import multiprocessing as mp
 import os
 import warnings
+from functools import partial
 from typing import Tuple, Union
 
 import matplotlib
@@ -10,11 +12,13 @@ import numpy as np
 import pandas
 import pandas as pd
 from tqdm import tqdm
+from tqdm.contrib.concurrent import process_map
 
 from ..gridding.QTree import QTree
 from ..gridding.QuadGrid import QuadGrid
 from .validation import check_transform_spatio_bin_jitter_magnitude, check_transform_temporal_bin_start_jitter
 
+# from ..multiprocessing.utils import load_data_from_shr_mem, load_data_to_shr_mem, mp_split_func_shr_mem
 os.environ["MKL_NUM_THREADS"] = "1"
 os.environ["NUMEXPR_NUM_THREADS"] = "1"
 os.environ["OMP_NUM_THREADS"] = "1"
@@ -67,6 +71,129 @@ def generate_temporal_bins(
         i += 1
 
     return bin_list
+
+
+def get_one_ensemble_quadtree(
+    ensemble_count,
+    size,
+    spatio_bin_jitter_magnitude,
+    temporal_start,
+    temporal_end,
+    temporal_step,
+    temporal_bin_interval,
+    temporal_bin_start_jitter,
+    data,
+    Temporal1,
+    grid_len,
+    grid_len_lon_upper_threshold,
+    grid_len_lon_lower_threshold,
+    grid_len_lat_upper_threshold,
+    grid_len_lat_lower_threshold,
+    points_lower_threshold,
+    plot_empty,
+    Spatio1,
+    Spatio2,
+    save_gridding_plot,
+    ax,
+):
+    rotation_angle = (90 / size) * ensemble_count
+    calibration_point_x_jitter = np.random.uniform(-spatio_bin_jitter_magnitude, spatio_bin_jitter_magnitude)
+    calibration_point_y_jitter = np.random.uniform(-spatio_bin_jitter_magnitude, spatio_bin_jitter_magnitude)
+
+    # print(f'ensemble_count: {ensemble_count}')
+    temporal_bins = generate_temporal_bins(
+        start=temporal_start,
+        end=temporal_end,
+        step=temporal_step,
+        bin_interval=temporal_bin_interval,
+        temporal_bin_start_jitter=temporal_bin_start_jitter,
+    )
+
+    ensemble_all_df_list = []
+
+    for time_block_index, bin_ in enumerate(temporal_bins):
+        time_start = bin_[0]
+        time_end = bin_[1]
+        sub_data = data[(data[Temporal1] >= time_start) & (data[Temporal1] < time_end)]
+
+        if len(sub_data) == 0:
+            continue
+
+        if grid_len is None:
+            QT_obj = QTree(
+                grid_len_lon_upper_threshold=grid_len_lon_upper_threshold,
+                grid_len_lon_lower_threshold=grid_len_lon_lower_threshold,
+                grid_len_lat_upper_threshold=grid_len_lat_upper_threshold,
+                grid_len_lat_lower_threshold=grid_len_lat_lower_threshold,
+                points_lower_threshold=points_lower_threshold,
+                lon_lat_equal_grid=True,
+                rotation_angle=rotation_angle,
+                calibration_point_x_jitter=calibration_point_x_jitter,
+                calibration_point_y_jitter=calibration_point_y_jitter,
+                plot_empty=plot_empty,
+            )
+        elif isinstance(grid_len, float) or isinstance(grid_len, int):
+            QT_obj = QuadGrid(
+                grid_len=grid_len,
+                points_lower_threshold=points_lower_threshold,
[…]
