+++ stemflow/utils/sphere_quadtree.py
@@ -2,8 +2,10 @@
 
 import os
 import warnings
+from functools import partial
 from typing import Tuple, Union
 
+import joblib
 import matplotlib
 import matplotlib.pyplot as plt  # plotting libraries
 import numpy as np
@@ -14,7 +16,6 @@ from tqdm import tqdm
 from ..gridding.Sphere_QTree import Sphere_QTree
 from .quadtree import generate_temporal_bins
 from .sphere.coordinate_transform import lonlat_cartesian_3D_transformer
-from .validation import check_njobs
 
 os.environ["MKL_NUM_THREADS"] = "1"
 os.environ["NUMEXPR_NUM_THREADS"] = "1"
@@ -23,10 +24,10 @@ os.environ["OMP_NUM_THREADS"] = "1"
 warnings.filterwarnings("ignore")
 
 
-def get_ensemble_sphere_quadtree(
+def get_one_ensemble_sphere_quadtree(
+    ensemble_count,
     data: pandas.core.frame.DataFrame,
     Temporal1: str = "DOY",
-    size: str = 1,
     grid_len_upper_threshold: Union[float, int] = 8000,
     grid_len_lower_threshold: Union[float, int] = 500,
     points_lower_threshold: int = 50,
@@ -38,15 +39,10 @@ def get_ensemble_sphere_quadtree(
     spatio_bin_jitter_magnitude: Union[float, int] = "adaptive",
     save_gridding_plotly: bool = True,
     save_gridding_plot: bool = False,
-    njobs: int = 1,
-    verbosity: int = 1,
-    plot_xlims: Tuple[Union[float, int]] = (-180, 180),
-    plot_ylims: Tuple[Union[float, int]] = (-90, 90),
-    save_path: str = "",
     ax=None,
     radius: Union[int, float] = 6371.0,
     plot_empty: bool = False,
-) -> Tuple[pandas.core.frame.DataFrame, Union[matplotlib.figure.Figure, float]]:
+):
     """Generate QuadTree gridding based on the input dataframe
     A function to get quadtree results for spherical indexing system. Twins to `get_ensemble_quadtree` in `quadtree.py`, Returns ensemble_df and plotting axes.
 
@@ -82,14 +78,6 @@ def get_ensemble_sphere_quadtree(
             Whether to save the plotly interactive gridding plot.
         save_gridding_plot:
             Whether ot save gridding plots
-        njobs:
-            Multi-processes count.
-        plot_xlims:
-            If save_gridding_plot=True, what is the xlims of the plot
-        plot_ylims:
-            If save_gridding_plot=True, what is the ylims of the plot
-        save_path:
-            If not '', save the ensemble dataframe to this path
         ax:
             Matplotlib Axes to add to.
         radius (Union[int, float]):
@@ -101,120 +89,72 @@ def get_ensemble_sphere_quadtree(
             2. grid plot. np.nan if save_gridding_plot=False<br>
 
     """
-    check_njobs(njobs)
+
+    if spatio_bin_jitter_magnitude == "adaptive":
+        rotation_angle = np.random.uniform(0, 90)
+        rotation_axis = np.random.uniform(-1, 1, 3)
+
+    temporal_bins = generate_temporal_bins(
+        start=temporal_start,
+        end=temporal_end,
+        step=temporal_step,
+        bin_interval=temporal_bin_interval,
+        temporal_bin_start_jitter=temporal_bin_start_jitter,
+    )
 
     ensemble_all_df_list = []
 
-    if save_gridding_plot:
-        if ax is None:
-            fig = plt.figure()
-            ax = fig.add_subplot(111, projection="3d")
-            ax.set_xlim([-radius, radius])
[…]
