+++ stemflow/model/AdaSTEM.py
@@ -10,6 +10,9 @@ from itertools import repeat
 from multiprocessing import Lock, Pool, Process, cpu_count, shared_memory
 from typing import Callable, Tuple, Union
 
+import joblib
+import matplotlib.pyplot as plt
+
 #
 # import dask.dataframe as dd
 import numpy as np
@@ -35,26 +38,18 @@ from sklearn.metrics import (
 from tqdm import tqdm
 from tqdm.auto import tqdm as tqdm_auto
 
-from ..multiprocessing.utils import (
-    load_data_from_shr_mem,
-    load_data_to_shr_mem,
-    load_model_from_shr_mem,
-    load_model_to_shr_mem,
-    mp_predict_func_shr_mem,
-    mp_training_func_shr_mem,
-)
-
 #
-from ..utils.quadtree import get_ensemble_quadtree
+from ..utils.quadtree import get_one_ensemble_quadtree
 from ..utils.validation import (
     check_base_model,
-    check_njobs,
     check_prediciton_aggregation,
     check_prediction_return,
     check_random_state,
     check_spatio_bin_jitter_magnitude,
     check_task,
     check_temporal_bin_start_jitter,
+    check_transform_njobs,
+    check_transform_spatio_bin_jitter_magnitude,
     check_verbosity,
     check_X_test,
     check_X_train,
@@ -249,7 +244,7 @@ class AdaSTEM(BaseEstimator):
         self.sample_weights_for_classifier = sample_weights_for_classifier
 
         # 5. Multi-threading params (not implemented yet)
-        check_njobs(njobs)
+        njobs = check_transform_njobs(self, njobs)
         self.njobs = njobs
 
         # 6. Plotting params
@@ -284,12 +279,12 @@ class AdaSTEM(BaseEstimator):
         Returns:
             self.grid_dict, a dictionary of one DataFrame for each grid, containing the gridding information
         """
-        # check_njobs(njobs)
+        njobs = check_transform_njobs(self, njobs)
 
         if verbosity is None:
             verbosity = self.verbosity
 
-        fold = self.ensemble_fold
+        # fold = self.ensemble_fold
         save_path = os.path.join(self.save_dir, "ensemble_quadtree_df.csv") if self.save_tmp else ""
 
         if "grid_len" not in self.__dir__():
@@ -299,33 +294,86 @@ class AdaSTEM(BaseEstimator):
             # We har using STEM
             pass
 
-        self.ensemble_df, self.gridding_plot = get_ensemble_quadtree(
-            X_train[[self.Spatio1, self.Spatio2, self.Temporal1]],
-            Spatio1=self.Spatio1,
-            Spatio2=self.Spatio2,
+        spatio_bin_jitter_magnitude = check_transform_spatio_bin_jitter_magnitude(
+            X_train, self.Spatio1, self.Spatio2, self.spatio_bin_jitter_magnitude
+        )
+
+        if self.save_gridding_plot:
+            if ax is None:
+                plt.figure(figsize=(20, 20))
+                plt.xlim([self.plot_xlims[0], self.plot_xlims[1]])
+                plt.ylim([self.plot_ylims[0], self.plot_ylims[1]])
+                plt.title("Quadtree", fontsize=20)
+            else:
+                pass
+
+        partial_get_one_ensemble_quadtree = partial(
+            get_one_ensemble_quadtree,
+            size=self.ensemble_fold,
+            spatio_bin_jitter_magnitude=spatio_bin_jitter_magnitude,
+            temporal_start=self.temporal_start,
+            temporal_end=self.temporal_end,
+            temporal_step=self.temporal_step,
+            temporal_bin_interval=self.temporal_bin_interval,
[…]

[…]
