+++ requirements.txt
@@ -1,3 +1,4 @@
+joblib>=1.3.0
 # geopandas>=0.11.1
 # h3pandas>=0.2.3
 matplotlib>=3.7.1
