### 2024-02-10 #6158 @earth-chris
hi @chenyangkang and @diazrenata, sorry for the delay but I'm just writing to confirm that I have checked the remaining boxes on my review. It looks like `jedalong` has too. Thanks for putting together a nice package, `chenyangkang`, and I look forward to working with `stemflow`.
