### 2024-02-15 #6158 @kthyng
I'm going to take over as editor!

Thanks for your review @earth-chris!

@jedalong I see your review list is checked off and the author opened some issues related to your comments which they have also subsequently closed. Can you look through the comments and issues to verify if your concerns have been addressed, and if you consider your review finished?
