### 2024-02-20 #6158 @jedalong
Hi yes the authors have made an attempt to address all of my comments and raised further concerns about leveraging spatial objects vs using more computational speedy operations, so this will always be a tradeoff. I consider my comments closed now.
Thanks
Jed
