### 2024-02-22 095a8cb
Commit along with new release 1.1 for completion of JOSS review
