### 2024-02-22 #6158 @chenyangkang
Hi @kthyng, I have completed the checklist. The new version is `v1.1`, releasing on figshare with DOI: 10.6084/m9.figshare.24669132
